Updated patient dashboard
