+++ src/Components/dashboard/Pages/Appointments.js
@@ -1,0 +1,191 @@
+import React from "react";
+
+export default function Appointments() {
+  return (
+    <div id="appointment">
+      <div className="row">
+        <div className="col-md-9">
+          <div className='row'>
+            <div className="col-md-4 col-sm-12 mb-2">
+            <div className="title_md">Appointments</div>
+            </div>
+            <div className="col-md-8 col-sm-12">
+            <div className="d-flex justify-content-end">
+              <div className="">
+                <button className="btn btn-appointment_1 mr-2">
+                  <svg
+                    width="25"
+                    height="25"
+                    viewBox="0 0 25 25"
+                    fill="none"
+                    xmlns="http://www.w3.org/2000/svg"
+                  >
+                    <path
+                      d="M22.125 3.5H20.25V2.75C20.25 2.35218 20.092 1.97064 19.8107 1.68934C19.5294 1.40804 19.1478 1.25 18.75 1.25C18.3522 1.25 17.9706 1.40804 17.6893 1.68934C17.408 1.97064 17.25 2.35218 17.25 2.75V3.5H14.25V2.75C14.25 2.35218 14.092 1.97064 13.8107 1.68934C13.5294 1.40804 13.1478 1.25 12.75 1.25C12.3522 1.25 11.9706 1.40804 11.6893 1.68934C11.408 1.97064 11.25 2.35218 11.25 2.75V3.5H8.25V2.75C8.25 2.35218 8.09196 1.97064 7.81066 1.68934C7.52936 1.40804 7.14782 1.25 6.75 1.25C6.35218 1.25 5.97064 1.40804 5.68934 1.68934C5.40804 1.97064 5.25 2.35218 5.25 2.75V3.5H3.375C2.87772 3.5 2.40081 3.69754 2.04917 4.04917C1.69754 4.40081 1.5 4.87772 1.5 5.375V20.375C1.5 20.8723 1.69754 21.3492 2.04917 21.7008C2.40081 22.0525 2.87772 22.25 3.375 22.25H13.875C13.9745 22.25 14.0698 22.2105 14.1402 22.1402C14.2105 22.0698 14.25 21.9745 14.25 21.875C14.25 21.7755 14.2105 21.6802 14.1402 21.6098C14.0698 21.5395 13.9745 21.5 13.875 21.5H3.375C3.07691 21.4991 2.79127 21.3803 2.58049 21.1695C2.3697 20.9587 2.25089 20.6731 2.25 20.375V8.75H23.25V15.125C23.25 15.2245 23.2895 15.3198 23.3598 15.3902C23.4302 15.4605 23.5255 15.5 23.625 15.5C23.7245 15.5 23.8198 15.4605 23.8902 15.3902C23.9605 15.3198 24 15.2245 24 15.125V5.375C24 4.87772 23.8025 4.40081 23.4508 4.04917C23.0992 3.69754 22.6223 3.5 22.125 3.5ZM18 2.75C18 2.55109 18.079 2.36032 18.2197 2.21967C18.3603 2.07902 18.5511 2 18.75 2C18.9489 2 19.1397 2.07902 19.2803 2.21967C19.421 2.36032 19.5 2.55109 19.5 2.75V5C19.5 5.19891 19.421 5.38968 19.2803 5.53033C19.1397 5.67098 18.9489 5.75 18.75 5.75C18.5511 5.75 18.3603 5.67098 18.2197 5.53033C18.079 5.38968 18 5.19891 18 5V2.75ZM12 2.75C12 2.55109 12.079 2.36032 12.2197 2.21967C12.3603 2.07902 12.5511 2 12.75 2C12.9489 2 13.1397 2.07902 13.2803 2.21967C13.421 2.36032 13.5 2.55109 13.5 2.75V5C13.5 5.19891 13.421 5.38968 13.2803 5.53033C13.1397 5.67098 12.9489 5.75 12.75 5.75C12.5511 5.75 12.3603 5.67098 12.2197 5.53033C12.079 5.38968 12 5.19891 12 5V2.75ZM6 2.75C6 2.55109 6.07902 2.36032 6.21967 2.21967C6.36032 2.07902 6.55109 2 6.75 2C6.94891 2 7.13968 2.07902 7.28033 2.21967C7.42098 2.36032 7.5 2.55109 7.5 2.75V5C7.5 5.19891 7.42098 5.38968 7.28033 5.53033C7.13968 5.67098 6.94891 5.75 6.75 5.75C6.55109 5.75 6.36032 5.67098 6.21967 5.53033C6.07902 5.38968 6 5.19891 6 5V2.75ZM23.25 8H2.25V5.375C2.25089 5.07691 2.3697 4.79127 2.58049 4.58049C2.79127 4.3697 3.07691 4.25089 3.375 4.25H5.25V5C5.25 5.39782 5.40804 5.77936 5.68934 6.06066C5.97064 6.34196 6.35218 6.5 6.75 6.5C7.14782 6.5 7.52936 6.34196 7.81066 6.06066C8.09196 5.77936 8.25 5.39782 8.25 5V4.25H11.25V5C11.25 5.39782 11.408 5.77936 11.6893 6.06066C11.9706 6.34196 12.3522 6.5 12.75 6.5C13.1478 6.5 13.5294 6.34196 13.8107 6.06066C14.092 5.77936 14.25 5.39782 14.25 5V4.25H17.25V5C17.25 5.39782 17.408 5.77936 17.6893 6.06066C17.9706 6.34196 18.3522 6.5 18.75 6.5C19.1478 6.5 19.5294 6.34196 19.8107 6.06066C20.092 5.77936 20.25 5.39782 20.25 5V4.25H22.125C22.4231 4.25089 22.7087 4.3697 22.9195 4.58049C23.1303 4.79127 23.2491 5.07691 23.25 5.375V8Z"
+                      fill="#141515"
+                    />
+                    <path
+                      d="M12 11.375C12 11.1761 11.921 10.9853 11.7803 10.8447C11.6397 10.704 11.4489 10.625 11.25 10.625H9.75C9.55109 10.625 9.36032 10.704 9.21967 10.8447C9.07902 10.9853 9 11.1761 9 11.375V12.5C9 12.6989 9.07902 12.8897 9.21967 13.0303C9.36032 13.171 9.55109 13.25 9.75 13.25H11.25C11.4489 13.25 11.6397 13.171 11.7803 13.0303C11.921 12.8897 12 12.6989 12 12.5V11.375ZM9.75 12.5V11.375H11.25V12.5H9.75Z"
+                      fill="#141515"
+                    />
+                    <path
+                      d="M7.5 11.375C7.5 11.1761 7.42098 10.9853 7.28033 10.8447C7.13968 10.704 6.94891 10.625 6.75 10.625H5.25C5.05109 10.625 4.86032 10.704 4.71967 10.8447C4.57902 10.9853 4.5 11.1761 4.5 11.375V12.5C4.5 12.6989 4.57902 12.8897 4.71967 13.0303C4.86032 13.171 5.05109 13.25 5.25 13.25H6.75C6.94891 13.25 7.13968 13.171 7.28033 13.0303C7.42098 12.8897 7.5 12.6989 7.5 12.5V11.375ZM5.25 12.5V11.375H6.75V12.5H5.25Z"
+                      fill="#141515"
+                    />
+                    <path
+                      d="M20.25 13.25C20.4489 13.25 20.6397 13.171 20.7803 13.0303C20.921 12.8897 21 12.6989 21 12.5V11.375C21 11.1761 20.921 10.9853 20.7803 10.8447C20.6397 10.704 20.4489 10.625 20.25 10.625H18.75C18.5511 10.625 18.3603 10.704 18.2197 10.8447C18.079 10.9853 18 11.1761 18 11.375V12.5C18 12.6989 18.079 12.8897 18.2197 13.0303C18.3603 13.171 18.5511 13.25 18.75 13.25H20.25ZM18.75 11.375H20.25V12.5H18.75V11.375Z"
+                      fill="#141515"
+                    />
+                    <path
+                      d="M12 14.75C12 14.5511 11.921 14.3603 11.7803 14.2197C11.6397 14.079 11.4489 14 11.25 14H9.75C9.55109 14 9.36032 14.079 9.21967 14.2197C9.07902 14.3603 9 14.5511 9 14.75V15.875C9 16.0739 9.07902 16.2647 9.21967 16.4053C9.36032 16.546 9.55109 16.625 9.75 16.625H11.25C11.4489 16.625 11.6397 16.546 11.7803 16.4053C11.921 16.2647 12 16.0739 12 15.875V14.75ZM9.75 15.875V14.75H11.25V15.875H9.75Z"
+                      fill="#141515"
+                    />
+                    <path
+                      d="M7.5 14.75C7.5 14.5511 7.42098 14.3603 7.28033 14.2197C7.13968 14.079 6.94891 14 6.75 14H5.25C5.05109 14 4.86032 14.079 4.71967 14.2197C4.57902 14.3603 4.5 14.5511 4.5 14.75V15.875C4.5 16.0739 4.57902 16.2647 4.71967 16.4053C4.86032 16.546 5.05109 16.625 5.25 16.625H6.75C6.94891 16.625 7.13968 16.546 7.28033 16.4053C7.42098 16.2647 7.5 16.0739 7.5 15.875V14.75ZM5.25 15.875V14.75H6.75V15.875H5.25Z"
+                      fill="#141515"
+                    />
+                    <path
+                      d="M11.25 17.375H9.75C9.55109 17.375 9.36032 17.454 9.21967 17.5947C9.07902 17.7353 9 17.9261 9 18.125V19.25C9 19.4489 9.07902 19.6397 9.21967 19.7803C9.36032 19.921 9.55109 20 9.75 20H11.25C11.4489 20 11.6397 19.921 11.7803 19.7803C11.921 19.6397 12 19.4489 12 19.25V18.125C12 17.9261 11.921 17.7353 11.7803 17.5947C11.6397 17.454 11.4489 17.375 11.25 17.375ZM9.75 19.25V18.125H11.25V19.25H9.75Z"
+                      fill="#141515"
+                    />
+                    <path
+                      d="M14.25 13.25H15.75C15.9489 13.25 16.1397 13.171 16.2803 13.0303C16.421 12.8897 16.5 12.6989 16.5 12.5V11.375C16.5 11.1761 16.421 10.9853 16.2803 10.8447C16.1397 10.704 15.9489 10.625 15.75 10.625H14.25C14.0511 10.625 13.8603 10.704 13.7197 10.8447C13.579 10.9853 13.5 11.1761 13.5 11.375V12.5C13.5 12.6989 13.579 12.8897 13.7197 13.0303C13.8603 13.171 14.0511 13.25 14.25 13.25ZM14.25 11.375H15.75V12.5H14.25V11.375Z"
+                      fill="#141515"
+                    />
+                    <path
+                      d="M13.5 15.875C13.5 16.0739 13.579 16.2647 13.7197 16.4053C13.8603 16.546 14.0511 16.625 14.25 16.625C14.3495 16.625 14.4448 16.5855 14.5152 16.5152C14.5855 16.4448 14.625 16.3495 14.625 16.25C14.625 16.1505 14.5855 16.0552 14.5152 15.9848C14.4448 15.9145 14.3495 15.875 14.25 15.875V14.75H15.75C15.8495 14.75 15.9448 14.7105 16.0152 14.6402C16.0855 14.5698 16.125 14.4745 16.125 14.375C16.125 14.2755 16.0855 14.1802 16.0152 14.1098C15.9448 14.0395 15.8495 14 15.75 14H14.25C14.0511 14 13.8603 14.079 13.7197 14.2197C13.579 14.3603 13.5 14.5511 13.5 14.75V15.875Z"
+                      fill="#141515"
+                    />
+                    <path
+                      d="M6.75 17.375H5.25C5.05109 17.375 4.86032 17.454 4.71967 17.5947C4.57902 17.7353 4.5 17.9261 4.5 18.125V19.25C4.5 19.4489 4.57902 19.6397 4.71967 19.7803C4.86032 19.921 5.05109 20 5.25 20H6.75C6.94891 20 7.13968 19.921 7.28033 19.7803C7.42098 19.6397 7.5 19.4489 7.5 19.25V18.125C7.5 17.9261 7.42098 17.7353 7.28033 17.5947C7.13968 17.454 6.94891 17.375 6.75 17.375ZM5.25 19.25V18.125H6.75V19.25H5.25Z"
+                      fill="#141515"
+                    />
+                    <path
+                      d="M19.125 14C18.1608 14 17.2183 14.2859 16.4166 14.8216C15.6149 15.3573 14.9901 16.1186 14.6211 17.0094C14.2521 17.9002 14.1556 18.8804 14.3437 19.8261C14.5318 20.7717 14.9961 21.6404 15.6779 22.3221C16.3596 23.0039 17.2283 23.4682 18.1739 23.6563C19.1196 23.8444 20.0998 23.7479 20.9906 23.3789C21.8814 23.0099 22.6427 22.3851 23.1784 21.5834C23.7141 20.7817 24 19.8392 24 18.875C23.9985 17.5825 23.4844 16.3434 22.5705 15.4295C21.6566 14.5156 20.4175 14.0015 19.125 14ZM19.125 23C18.3091 23 17.5116 22.7581 16.8333 22.3048C16.1549 21.8516 15.6262 21.2073 15.314 20.4536C15.0018 19.6998 14.9201 18.8704 15.0793 18.0703C15.2384 17.2701 15.6313 16.5351 16.2082 15.9582C16.7851 15.3813 17.5201 14.9884 18.3202 14.8293C19.1204 14.6701 19.9498 14.7518 20.7036 15.064C21.4573 15.3762 22.1015 15.9049 22.5548 16.5833C23.0081 17.2616 23.25 18.0592 23.25 18.875C23.2487 19.9686 22.8137 21.0171 22.0404 21.7904C21.2671 22.5637 20.2186 22.9987 19.125 23Z"
+                      fill="#141515"
+                    />
+                    <path
+                      d="M21.2919 17.303L18.3752 20.2197L17.0496 18.8937C16.9789 18.8254 16.8841 18.7876 16.7858 18.7885C16.6875 18.7894 16.5934 18.8288 16.5239 18.8983C16.4544 18.9678 16.4149 19.0619 16.4141 19.1602C16.4132 19.2585 16.451 19.3533 16.5193 19.424L18.1101 21.0151C18.1804 21.0854 18.2758 21.1249 18.3752 21.1249C18.4746 21.1249 18.57 21.0854 18.6403 21.0151L21.8222 17.8333C21.8905 17.7625 21.9283 17.6678 21.9275 17.5695C21.9266 17.4712 21.8872 17.3771 21.8176 17.3076C21.7481 17.238 21.654 17.1986 21.5557 17.1977C21.4574 17.1969 21.3627 17.2347 21.2919 17.303Z"
+                      fill="#141515"
+                    />
+                  </svg>
+                  <span className="ml-2">Aug 1 - Sept 4, 2020</span>
+                </button>
+              </div>
+
+              <div className="">
+                <button className="btn btn-appointment">
+                  <svg
+                    width="25"
+                    height="24"
+                    viewBox="0 0 25 24"
+                    fill="none"
+                    xmlns="http://www.w3.org/2000/svg"
+                  >
+                    <path
+                      d="M12.75 5V19"
+                      stroke="white"
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                    />
+                    <path
+                      d="M5.75 12H19.75"
+                      stroke="white"
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                    />
+                  </svg>
+                  <span className="">New Appointment</span>
+                </button>
+              </div>
+            </div>
+            </div>
+          </div>
+          <div className="table-responsive mt-4">
+          <table class="table" id="appointment">
+            <thead>
+              <tr>
+                <th scope="col">Thur Aug 27</th>
+                <th scope="col">Fri Aug 28</th>
+                <th scope="col">Sat Aug 29</th>
+                <th scope="col">Sun Aug 30</th>
+                <th scope="col">Mon Aug 31</th>
+              </tr>
+            </thead>
+            <tbody>
+              <tr>
+                <th scope="row">1</th>
+                <td>Mark</td>
+                <td>Otto</td>
+                <td>@mdo</td>
+              </tr>
+              <tr>
+                <th scope="row">2</th>
+                <td>Jacob</td>
+                <td>Thornton</td>
+                <td>@fat</td>
+              </tr>
+              <tr>
+                <th scope="row">3</th>
+                <td>Larry</td>
+                <td>the Bird</td>
+                <td>@twitter</td>
+              </tr>
+            </tbody>
+          </table>
+          </div>
+
+        </div>
+        <div className="col-md-3">
+          <div className="vl">
+            <p className="mb-0 title_5">Doctors</p>
+            <div class="search-container mt-2 mb-4">
+
+              <form class="form-group d-flex justify-content-between">
+                <input
+                  class="form-control"
+                  type="search"
+                  placeholder="Search for doctor or nurse"
+                  aria-label="Search"
+                />
+                <div class="search-icon py-1">
+                  <svg
+                    width="25"
+                    height="24"
+                    viewBox="0 0 25 24"
+                    fill="none"
+                    xmlns="http://www.w3.org/2000/svg"
+                  >
+                    <path
+                      d="M11.0977 19C15.5159 19 19.0977 15.4183 19.0977 11C19.0977 6.58172 15.5159 3 11.0977 3C6.67938 3 3.09766 6.58172 3.09766 11C3.09766 15.4183 6.67938 19 11.0977 19Z"
+                      stroke="#706868"
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                    />
+                    <path
+                      d="M21.098 21L16.748 16.65"
+                      stroke="#706868"
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                    />
+                  </svg>
+                </div>
+              </form>
+
+
+            </div>
+            
+            <div>
+              <ul>
+                <li className="title_1 d-flex justify-content-between my-2"><span>Dr Abola Jimoh</span><span className="d3ext">A</span></li>
+                <li className="title_1 d-flex justify-content-between my-2"><span>Dr Abola Jimoh</span><span className="d3ext">A</span></li>
+                <li className="title_1 d-flex justify-content-between my-2"><span>Dr Abola Jimoh</span><span className="d3ext">A</span></li>
+              </ul>
+
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
