+++ src/Components/dashboard/Pages/DashboardHospital.js
@@ -1,0 +1,274 @@
+import React, { Component } from "react";
+
+export default class DashboardHospital extends Component {
+  render() {
+    return (
+      <div>
+        <div className="row">
+          <div className="col-md-8 col-sm-12">
+            <div className="card mb-4">
+              <div className="card-header">
+                <div class="d-flex justify-content-between">
+                  <div>Appointments</div>
+                  <svg
+                    width="25"
+                    height="24"
+                    viewBox="0 0 25 24"
+                    fill="none"
+                    xmlns="http://www.w3.org/2000/svg"
+                  >
+                    <g clip-path="url(#clip0)">
+                      <path
+                        d="M21.4688 3H19.5938V2.25C19.5938 1.85218 19.4357 1.47064 19.1544 1.18934C18.8731 0.908035 18.4916 0.75 18.0938 0.75C17.6959 0.75 17.3144 0.908035 17.0331 1.18934C16.7518 1.47064 16.5938 1.85218 16.5938 2.25V3H13.5938V2.25C13.5938 1.85218 13.4357 1.47064 13.1544 1.18934C12.8731 0.908035 12.4916 0.75 12.0938 0.75C11.6959 0.75 11.3144 0.908035 11.0331 1.18934C10.7518 1.47064 10.5938 1.85218 10.5938 2.25V3H7.59375V2.25C7.59375 1.85218 7.43571 1.47064 7.15441 1.18934C6.87311 0.908035 6.49157 0.75 6.09375 0.75C5.69593 0.75 5.31439 0.908035 5.03309 1.18934C4.75179 1.47064 4.59375 1.85218 4.59375 2.25V3H2.71875C2.22147 3 1.74456 3.19754 1.39292 3.54917C1.04129 3.90081 0.84375 4.37772 0.84375 4.875V19.875C0.84375 20.3723 1.04129 20.8492 1.39292 21.2008C1.74456 21.5525 2.22147 21.75 2.71875 21.75H13.2188C13.3182 21.75 13.4136 21.7105 13.4839 21.6402C13.5542 21.5698 13.5938 21.4745 13.5938 21.375C13.5938 21.2755 13.5542 21.1802 13.4839 21.1098C13.4136 21.0395 13.3182 21 13.2188 21H2.71875C2.42066 20.9991 2.13502 20.8803 1.92424 20.6695C1.71345 20.4587 1.59464 20.1731 1.59375 19.875V8.25H22.5938V14.625C22.5938 14.7245 22.6333 14.8198 22.7036 14.8902C22.7739 14.9605 22.8693 15 22.9688 15C23.0682 15 23.1636 14.9605 23.2339 14.8902C23.3042 14.8198 23.3438 14.7245 23.3438 14.625V4.875C23.3438 4.37772 23.1462 3.90081 22.7946 3.54917C22.4429 3.19754 21.966 3 21.4688 3ZM17.3438 2.25C17.3438 2.05109 17.4228 1.86032 17.5634 1.71967C17.7041 1.57902 17.8948 1.5 18.0938 1.5C18.2927 1.5 18.4834 1.57902 18.6241 1.71967C18.7647 1.86032 18.8438 2.05109 18.8438 2.25V4.5C18.8438 4.69891 18.7647 4.88968 18.6241 5.03033C18.4834 5.17098 18.2927 5.25 18.0938 5.25C17.8948 5.25 17.7041 5.17098 17.5634 5.03033C17.4228 4.88968 17.3438 4.69891 17.3438 4.5V2.25ZM11.3438 2.25C11.3438 2.05109 11.4228 1.86032 11.5634 1.71967C11.7041 1.57902 11.8948 1.5 12.0938 1.5C12.2927 1.5 12.4834 1.57902 12.6241 1.71967C12.7647 1.86032 12.8438 2.05109 12.8438 2.25V4.5C12.8438 4.69891 12.7647 4.88968 12.6241 5.03033C12.4834 5.17098 12.2927 5.25 12.0938 5.25C11.8948 5.25 11.7041 5.17098 11.5634 5.03033C11.4228 4.88968 11.3438 4.69891 11.3438 4.5V2.25ZM5.34375 2.25C5.34375 2.05109 5.42277 1.86032 5.56342 1.71967C5.70407 1.57902 5.89484 1.5 6.09375 1.5C6.29266 1.5 6.48343 1.57902 6.62408 1.71967C6.76473 1.86032 6.84375 2.05109 6.84375 2.25V4.5C6.84375 4.69891 6.76473 4.88968 6.62408 5.03033C6.48343 5.17098 6.29266 5.25 6.09375 5.25C5.89484 5.25 5.70407 5.17098 5.56342 5.03033C5.42277 4.88968 5.34375 4.69891 5.34375 4.5V2.25ZM22.5938 7.5H1.59375V4.875C1.59464 4.57691 1.71345 4.29127 1.92424 4.08049C2.13502 3.8697 2.42066 3.75089 2.71875 3.75H4.59375V4.5C4.59375 4.89782 4.75179 5.27936 5.03309 5.56066C5.31439 5.84196 5.69593 6 6.09375 6C6.49157 6 6.87311 5.84196 7.15441 5.56066C7.43571 5.27936 7.59375 4.89782 7.59375 4.5V3.75H10.5938V4.5C10.5938 4.89782 10.7518 5.27936 11.0331 5.56066C11.3144 5.84196 11.6959 6 12.0938 6C12.4916 6 12.8731 5.84196 13.1544 5.56066C13.4357 5.27936 13.5938 4.89782 13.5938 4.5V3.75H16.5938V4.5C16.5938 4.89782 16.7518 5.27936 17.0331 5.56066C17.3144 5.84196 17.6959 6 18.0938 6C18.4916 6 18.8731 5.84196 19.1544 5.56066C19.4357 5.27936 19.5938 4.89782 19.5938 4.5V3.75H21.4688C21.7668 3.75089 22.0525 3.8697 22.2633 4.08049C22.474 4.29127 22.5929 4.57691 22.5938 4.875V7.5Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M11.3438 10.875C11.3438 10.6761 11.2647 10.4853 11.1241 10.3447C10.9834 10.204 10.7927 10.125 10.5938 10.125H9.09375C8.89484 10.125 8.70407 10.204 8.56342 10.3447C8.42277 10.4853 8.34375 10.6761 8.34375 10.875V12C8.34375 12.1989 8.42277 12.3897 8.56342 12.5303C8.70407 12.671 8.89484 12.75 9.09375 12.75H10.5938C10.7927 12.75 10.9834 12.671 11.1241 12.5303C11.2647 12.3897 11.3438 12.1989 11.3438 12V10.875ZM9.09375 12V10.875H10.5938V12H9.09375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M6.84375 10.875C6.84375 10.6761 6.76473 10.4853 6.62408 10.3447C6.48343 10.204 6.29266 10.125 6.09375 10.125H4.59375C4.39484 10.125 4.20407 10.204 4.06342 10.3447C3.92277 10.4853 3.84375 10.6761 3.84375 10.875V12C3.84375 12.1989 3.92277 12.3897 4.06342 12.5303C4.20407 12.671 4.39484 12.75 4.59375 12.75H6.09375C6.29266 12.75 6.48343 12.671 6.62408 12.5303C6.76473 12.3897 6.84375 12.1989 6.84375 12V10.875ZM4.59375 12V10.875H6.09375V12H4.59375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M19.5938 12.75C19.7927 12.75 19.9834 12.671 20.1241 12.5303C20.2647 12.3897 20.3438 12.1989 20.3438 12V10.875C20.3438 10.6761 20.2647 10.4853 20.1241 10.3447C19.9834 10.204 19.7927 10.125 19.5938 10.125H18.0938C17.8948 10.125 17.7041 10.204 17.5634 10.3447C17.4228 10.4853 17.3438 10.6761 17.3438 10.875V12C17.3438 12.1989 17.4228 12.3897 17.5634 12.5303C17.7041 12.671 17.8948 12.75 18.0938 12.75H19.5938ZM18.0938 10.875H19.5938V12H18.0938V10.875Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M11.3438 14.25C11.3438 14.0511 11.2647 13.8603 11.1241 13.7197C10.9834 13.579 10.7927 13.5 10.5938 13.5H9.09375C8.89484 13.5 8.70407 13.579 8.56342 13.7197C8.42277 13.8603 8.34375 14.0511 8.34375 14.25V15.375C8.34375 15.5739 8.42277 15.7647 8.56342 15.9053C8.70407 16.046 8.89484 16.125 9.09375 16.125H10.5938C10.7927 16.125 10.9834 16.046 11.1241 15.9053C11.2647 15.7647 11.3438 15.5739 11.3438 15.375V14.25ZM9.09375 15.375V14.25H10.5938V15.375H9.09375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M6.84375 14.25C6.84375 14.0511 6.76473 13.8603 6.62408 13.7197C6.48343 13.579 6.29266 13.5 6.09375 13.5H4.59375C4.39484 13.5 4.20407 13.579 4.06342 13.7197C3.92277 13.8603 3.84375 14.0511 3.84375 14.25V15.375C3.84375 15.5739 3.92277 15.7647 4.06342 15.9053C4.20407 16.046 4.39484 16.125 4.59375 16.125H6.09375C6.29266 16.125 6.48343 16.046 6.62408 15.9053C6.76473 15.7647 6.84375 15.5739 6.84375 15.375V14.25ZM4.59375 15.375V14.25H6.09375V15.375H4.59375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M10.5938 16.875H9.09375C8.89484 16.875 8.70407 16.954 8.56342 17.0947C8.42277 17.2353 8.34375 17.4261 8.34375 17.625V18.75C8.34375 18.9489 8.42277 19.1397 8.56342 19.2803C8.70407 19.421 8.89484 19.5 9.09375 19.5H10.5938C10.7927 19.5 10.9834 19.421 11.1241 19.2803C11.2647 19.1397 11.3438 18.9489 11.3438 18.75V17.625C11.3438 17.4261 11.2647 17.2353 11.1241 17.0947C10.9834 16.954 10.7927 16.875 10.5938 16.875ZM9.09375 18.75V17.625H10.5938V18.75H9.09375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M13.5938 12.75H15.0938C15.2927 12.75 15.4834 12.671 15.6241 12.5303C15.7647 12.3897 15.8438 12.1989 15.8438 12V10.875C15.8438 10.6761 15.7647 10.4853 15.6241 10.3447C15.4834 10.204 15.2927 10.125 15.0938 10.125H13.5938C13.3948 10.125 13.2041 10.204 13.0634 10.3447C12.9228 10.4853 12.8438 10.6761 12.8438 10.875V12C12.8438 12.1989 12.9228 12.3897 13.0634 12.5303C13.2041 12.671 13.3948 12.75 13.5938 12.75ZM13.5938 10.875H15.0938V12H13.5938V10.875Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M12.8438 15.375C12.8438 15.5739 12.9228 15.7647 13.0634 15.9053C13.2041 16.046 13.3948 16.125 13.5938 16.125C13.6932 16.125 13.7886 16.0855 13.8589 16.0152C13.9292 15.9448 13.9688 15.8495 13.9688 15.75C13.9688 15.6505 13.9292 15.5552 13.8589 15.4848C13.7886 15.4145 13.6932 15.375 13.5938 15.375V14.25H15.0938C15.1932 14.25 15.2886 14.2105 15.3589 14.1402C15.4292 14.0698 15.4688 13.9745 15.4688 13.875C15.4688 13.7755 15.4292 13.6802 15.3589 13.6098C15.2886 13.5395 15.1932 13.5 15.0938 13.5H13.5938C13.3948 13.5 13.2041 13.579 13.0634 13.7197C12.9228 13.8603 12.8438 14.0511 12.8438 14.25V15.375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M6.09375 16.875H4.59375C4.39484 16.875 4.20407 16.954 4.06342 17.0947C3.92277 17.2353 3.84375 17.4261 3.84375 17.625V18.75C3.84375 18.9489 3.92277 19.1397 4.06342 19.2803C4.20407 19.421 4.39484 19.5 4.59375 19.5H6.09375C6.29266 19.5 6.48343 19.421 6.62408 19.2803C6.76473 19.1397 6.84375 18.9489 6.84375 18.75V17.625C6.84375 17.4261 6.76473 17.2353 6.62408 17.0947C6.48343 16.954 6.29266 16.875 6.09375 16.875ZM4.59375 18.75V17.625H6.09375V18.75H4.59375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M18.4687 13.5C17.5046 13.5 16.562 13.7859 15.7603 14.3216C14.9587 14.8573 14.3338 15.6186 13.9648 16.5094C13.5959 17.4002 13.4993 18.3804 13.6874 19.3261C13.8755 20.2717 14.3398 21.1404 15.0216 21.8221C15.7034 22.5039 16.572 22.9682 17.5177 23.1563C18.4633 23.3444 19.4435 23.2479 20.3343 22.8789C21.2251 22.5099 21.9865 21.8851 22.5222 21.0834C23.0578 20.2817 23.3437 19.3392 23.3437 18.375C23.3422 17.0825 22.8282 15.8434 21.9142 14.9295C21.0003 14.0156 19.7612 13.5015 18.4687 13.5ZM18.4687 22.5C17.6529 22.5 16.8554 22.2581 16.177 21.8048C15.4987 21.3516 14.97 20.7073 14.6577 19.9536C14.3455 19.1998 14.2638 18.3704 14.423 17.5703C14.5822 16.7701 14.975 16.0351 15.5519 15.4582C16.1288 14.8813 16.8638 14.4884 17.664 14.3293C18.4642 14.1701 19.2936 14.2518 20.0473 14.564C20.8011 14.8762 21.4453 15.4049 21.8986 16.0833C22.3518 16.7616 22.5937 17.5592 22.5937 18.375C22.5924 19.4686 22.1574 20.5171 21.3841 21.2904C20.6108 22.0637 19.5624 22.4987 18.4687 22.5Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M20.6357 16.803L17.719 19.7198L16.3933 18.3938C16.3226 18.3254 16.2279 18.2876 16.1296 18.2885C16.0312 18.2894 15.9372 18.3288 15.8676 18.3983C15.7981 18.4679 15.7587 18.5619 15.7578 18.6602C15.757 18.7586 15.7948 18.8533 15.8631 18.924L17.4538 20.5151C17.5241 20.5854 17.6195 20.6249 17.719 20.6249C17.8184 20.6249 17.9138 20.5854 17.9841 20.5151L21.1659 17.3333C21.2343 17.2625 21.2721 17.1678 21.2712 17.0695C21.2703 16.9712 21.2309 16.8771 21.1614 16.8076C21.0919 16.738 20.9978 16.6986 20.8995 16.6978C20.8011 16.6969 20.7064 16.7347 20.6357 16.803Z"
+                        fill="#FB5B45"
+                      />
+                    </g>
+                    <defs>
+                      <clipPath id="clip0">
+                        <rect
+                          width="24"
+                          height="24"
+                          fill="white"
+                          transform="translate(0.09375)"
+                        />
+                      </clipPath>
+                    </defs>
+                  </svg>
+                </div>
+              </div>
+              <div className="card-body">
+                <div class="row mb-2">
+                  <div class="col-md-3">
+                    <div class="d-flex mb-3">
+                      <div>
+                        <img
+                          class="rounded-img-sm mr-2"
+                          src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                        />
+                      </div>
+                      <div class="">
+                        <p class="m-0 p-0 title_1">Aboki Joke</p>
+                        <p class="m-0 p-0 title_2">03 Aug, 2020</p>
+                      </div>
+                    </div>
+                  </div>
+                  <div class="col-md-9">
+                    <div class="">
+                      <div class="m-0 title_2">
+                        Dr Elliot updated your medical records on Friday 4th
+                        August, 2020 based on your complaints and diagnosis.
+                      </div>
+                    </div>
+                  </div>
+                </div>
+
+                <div class="row mb-2">
+                  <div class="col-md-3">
+                    <div class="d-flex mb-3">
+                      <div>
+                        <img
+                          class="rounded-img-sm mr-2"
+                          src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                        />
+                      </div>
+                      <div class="">
+                        <p class="m-0 p-0 title_1">Aboki Joke</p>
+                        <p class="m-0 p-0 title_2">03 Aug, 2020</p>
+                      </div>
+                    </div>
+                  </div>
+                  <div class="col-md-9">
+                    <div class="">
+                      <div class="m-0 title_2">
+                        Dr Elliot updated your medical records on Friday 4th
+                        August, 2020 based on your complaints and diagnosis.
+                      </div>
+                    </div>
+                  </div>
+                </div>
+
+                <div class="row mb-2">
+                  <div class="col-md-3">
+                    <div class="d-flex mb-3">
+                      <div>
+                        <img
+                          class="rounded-img-sm mr-2"
+                          src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                        />
+                      </div>
+                      <div class="">
+                        <p class="m-0 p-0 title_1">Aboki Joke</p>
+                        <p class="m-0 p-0 title_2">03 Aug, 2020</p>
+                      </div>
+                    </div>
+                  </div>
+                  <div class="col-md-9">
+                    <div class="">
+                      <div class="m-0 title_2">
+                        Dr Elliot updated your medical records on Friday 4th
+                        August, 2020 based on your complaints and diagnosis.
+                      </div>
+                    </div>
+                  </div>
+                </div>
+
+                <div class="row mb-2">
+                  <div class="col-md-3">
+                    <div class="d-flex mb-3">
+                      <div>
+                        <img
+                          class="rounded-img-sm mr-2"
+                          src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                        />
+                      </div>
+                      <div class="">
+                        <p class="m-0 p-0 title_1">Aboki Joke</p>
+                        <p class="m-0 p-0 title_2">03 Aug, 2020</p>
+                      </div>
+                    </div>
+                  </div>
+                  <div class="col-md-9">
+                    <div class="">
+                      <div class="m-0 title_2">
+                        Dr Elliot updated your medical records on Friday 4th
+                        August, 2020 based on your complaints and diagnosis.
+                      </div>
+                    </div>
+                  </div>
+                </div>
+              </div>
+              <div class="card-footer text-muted">
+                <div class="title text-right">See All Doctors</div>
+              </div>
+            </div>
+          </div>
+          <div className="col-md-4 col-sm-12">
+            <div>
+              <div class="card mb-4">
+              <div className="card-header">
+                  <div class="heading-title">
+                    Emergency Calls
+                  </div>
+                </div>
+                <div class="card-body">
+                  <div class="py-3 b_b">
+                    <div class="d-flex mb-1">
+                      <div>
+                        <img
+                          class="rounded-img-sm mr-2"
+                          src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                        />
+                      </div>
+                      <div class="d-block">
+                        <p class="m-0 p-0 title_1">A. Jhumla</p>
+                        <p class="m-0 p-0 title_2">Need an ambulance</p>
+                      </div>
+                    </div>
+                    <div class="title_2">03 Aug; 3:10 PM</div>
+                    <div class="title_8 text-center mt-3 m-1">
+                      A-45, amdrew street, fmur astm.
+                    </div>
+                  </div>
+                  <div class="py-3 b_b">
+                    <div class="d-flex mb-1">
+                      <div>
+                        <img
+                          class="rounded-img-sm mr-2"
+                          src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                        />
+                      </div>
+                      <div class="d-block">
+                        <p class="m-0 p-0 title_1">Dr Emeka Emeka Ike</p>
+                        <p class="m-0 p-0 title_2">General Surgeon</p>
+                      </div>
+                    </div>
+                    <div class="title_2">03 Aug; 3:10 PM</div>
+                    <div class="title_8 text-center mt-3 m-1">
+                      Schedule Appointment
+                    </div>
+                  </div>
+                </div>
+                <div class="card-footer">
+                  <div class="title text-right">See All</div>
+                </div>
+              </div>
+
+              <div className="card my-2">
+                <div className="card-header">
+                  <div class="heading-title">
+                    Patient Appointments
+                  </div>
+                </div>
+                <div className="card-body">
+                  <div class="d-flex mb-3">
+                    <div>
+                      <img
+                        class="rounded-img-sm mr-2"
+                        src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                      />
+                    </div>
+                    <div class="">
+                      <p class="m-0 p-0 title_1">Dr Shola Jimoh</p>
+                      <p class="m-0 p-0 title_2">
+                        Lorem Ipsum Lorem Ipsum Lorem Ipsum Lorem Ipsum Lorem
+                        Ipsum Lorem Ipsum Lorem Ipsum
+                      </p>
+                    </div>
+                  </div>
+                </div>
+                <div className="card-footer d-flex justify-content-between">
+                  <div class="title">Confirm and Appoint Doctor</div>
+                  <div class="title">See All</div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    );
+  }
+}
