+++ src/Components/dashboard/Pages/DashboardPatient.js
@@ -1,0 +1,603 @@
+import React, { Component } from "react";
+
+export default class DashboardPatient extends Component {
+  render() {
+    return (
+      <div>
+        <div class="row">
+          <div class="col-md-4 mb-2">
+            <div class="card">
+              <div class="card-header">
+                <div class="d-flex justify-content-between">
+                  <div>Appointments</div>
+                  <svg
+                    width="25"
+                    height="24"
+                    viewBox="0 0 25 24"
+                    fill="none"
+                    xmlns="http://www.w3.org/2000/svg"
+                  >
+                    <g clip-path="url(#clip0)">
+                      <path
+                        d="M21.4688 3H19.5938V2.25C19.5938 1.85218 19.4357 1.47064 19.1544 1.18934C18.8731 0.908035 18.4916 0.75 18.0938 0.75C17.6959 0.75 17.3144 0.908035 17.0331 1.18934C16.7518 1.47064 16.5938 1.85218 16.5938 2.25V3H13.5938V2.25C13.5938 1.85218 13.4357 1.47064 13.1544 1.18934C12.8731 0.908035 12.4916 0.75 12.0938 0.75C11.6959 0.75 11.3144 0.908035 11.0331 1.18934C10.7518 1.47064 10.5938 1.85218 10.5938 2.25V3H7.59375V2.25C7.59375 1.85218 7.43571 1.47064 7.15441 1.18934C6.87311 0.908035 6.49157 0.75 6.09375 0.75C5.69593 0.75 5.31439 0.908035 5.03309 1.18934C4.75179 1.47064 4.59375 1.85218 4.59375 2.25V3H2.71875C2.22147 3 1.74456 3.19754 1.39292 3.54917C1.04129 3.90081 0.84375 4.37772 0.84375 4.875V19.875C0.84375 20.3723 1.04129 20.8492 1.39292 21.2008C1.74456 21.5525 2.22147 21.75 2.71875 21.75H13.2188C13.3182 21.75 13.4136 21.7105 13.4839 21.6402C13.5542 21.5698 13.5938 21.4745 13.5938 21.375C13.5938 21.2755 13.5542 21.1802 13.4839 21.1098C13.4136 21.0395 13.3182 21 13.2188 21H2.71875C2.42066 20.9991 2.13502 20.8803 1.92424 20.6695C1.71345 20.4587 1.59464 20.1731 1.59375 19.875V8.25H22.5938V14.625C22.5938 14.7245 22.6333 14.8198 22.7036 14.8902C22.7739 14.9605 22.8693 15 22.9688 15C23.0682 15 23.1636 14.9605 23.2339 14.8902C23.3042 14.8198 23.3438 14.7245 23.3438 14.625V4.875C23.3438 4.37772 23.1462 3.90081 22.7946 3.54917C22.4429 3.19754 21.966 3 21.4688 3ZM17.3438 2.25C17.3438 2.05109 17.4228 1.86032 17.5634 1.71967C17.7041 1.57902 17.8948 1.5 18.0938 1.5C18.2927 1.5 18.4834 1.57902 18.6241 1.71967C18.7647 1.86032 18.8438 2.05109 18.8438 2.25V4.5C18.8438 4.69891 18.7647 4.88968 18.6241 5.03033C18.4834 5.17098 18.2927 5.25 18.0938 5.25C17.8948 5.25 17.7041 5.17098 17.5634 5.03033C17.4228 4.88968 17.3438 4.69891 17.3438 4.5V2.25ZM11.3438 2.25C11.3438 2.05109 11.4228 1.86032 11.5634 1.71967C11.7041 1.57902 11.8948 1.5 12.0938 1.5C12.2927 1.5 12.4834 1.57902 12.6241 1.71967C12.7647 1.86032 12.8438 2.05109 12.8438 2.25V4.5C12.8438 4.69891 12.7647 4.88968 12.6241 5.03033C12.4834 5.17098 12.2927 5.25 12.0938 5.25C11.8948 5.25 11.7041 5.17098 11.5634 5.03033C11.4228 4.88968 11.3438 4.69891 11.3438 4.5V2.25ZM5.34375 2.25C5.34375 2.05109 5.42277 1.86032 5.56342 1.71967C5.70407 1.57902 5.89484 1.5 6.09375 1.5C6.29266 1.5 6.48343 1.57902 6.62408 1.71967C6.76473 1.86032 6.84375 2.05109 6.84375 2.25V4.5C6.84375 4.69891 6.76473 4.88968 6.62408 5.03033C6.48343 5.17098 6.29266 5.25 6.09375 5.25C5.89484 5.25 5.70407 5.17098 5.56342 5.03033C5.42277 4.88968 5.34375 4.69891 5.34375 4.5V2.25ZM22.5938 7.5H1.59375V4.875C1.59464 4.57691 1.71345 4.29127 1.92424 4.08049C2.13502 3.8697 2.42066 3.75089 2.71875 3.75H4.59375V4.5C4.59375 4.89782 4.75179 5.27936 5.03309 5.56066C5.31439 5.84196 5.69593 6 6.09375 6C6.49157 6 6.87311 5.84196 7.15441 5.56066C7.43571 5.27936 7.59375 4.89782 7.59375 4.5V3.75H10.5938V4.5C10.5938 4.89782 10.7518 5.27936 11.0331 5.56066C11.3144 5.84196 11.6959 6 12.0938 6C12.4916 6 12.8731 5.84196 13.1544 5.56066C13.4357 5.27936 13.5938 4.89782 13.5938 4.5V3.75H16.5938V4.5C16.5938 4.89782 16.7518 5.27936 17.0331 5.56066C17.3144 5.84196 17.6959 6 18.0938 6C18.4916 6 18.8731 5.84196 19.1544 5.56066C19.4357 5.27936 19.5938 4.89782 19.5938 4.5V3.75H21.4688C21.7668 3.75089 22.0525 3.8697 22.2633 4.08049C22.474 4.29127 22.5929 4.57691 22.5938 4.875V7.5Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M11.3438 10.875C11.3438 10.6761 11.2647 10.4853 11.1241 10.3447C10.9834 10.204 10.7927 10.125 10.5938 10.125H9.09375C8.89484 10.125 8.70407 10.204 8.56342 10.3447C8.42277 10.4853 8.34375 10.6761 8.34375 10.875V12C8.34375 12.1989 8.42277 12.3897 8.56342 12.5303C8.70407 12.671 8.89484 12.75 9.09375 12.75H10.5938C10.7927 12.75 10.9834 12.671 11.1241 12.5303C11.2647 12.3897 11.3438 12.1989 11.3438 12V10.875ZM9.09375 12V10.875H10.5938V12H9.09375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M6.84375 10.875C6.84375 10.6761 6.76473 10.4853 6.62408 10.3447C6.48343 10.204 6.29266 10.125 6.09375 10.125H4.59375C4.39484 10.125 4.20407 10.204 4.06342 10.3447C3.92277 10.4853 3.84375 10.6761 3.84375 10.875V12C3.84375 12.1989 3.92277 12.3897 4.06342 12.5303C4.20407 12.671 4.39484 12.75 4.59375 12.75H6.09375C6.29266 12.75 6.48343 12.671 6.62408 12.5303C6.76473 12.3897 6.84375 12.1989 6.84375 12V10.875ZM4.59375 12V10.875H6.09375V12H4.59375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M19.5938 12.75C19.7927 12.75 19.9834 12.671 20.1241 12.5303C20.2647 12.3897 20.3438 12.1989 20.3438 12V10.875C20.3438 10.6761 20.2647 10.4853 20.1241 10.3447C19.9834 10.204 19.7927 10.125 19.5938 10.125H18.0938C17.8948 10.125 17.7041 10.204 17.5634 10.3447C17.4228 10.4853 17.3438 10.6761 17.3438 10.875V12C17.3438 12.1989 17.4228 12.3897 17.5634 12.5303C17.7041 12.671 17.8948 12.75 18.0938 12.75H19.5938ZM18.0938 10.875H19.5938V12H18.0938V10.875Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M11.3438 14.25C11.3438 14.0511 11.2647 13.8603 11.1241 13.7197C10.9834 13.579 10.7927 13.5 10.5938 13.5H9.09375C8.89484 13.5 8.70407 13.579 8.56342 13.7197C8.42277 13.8603 8.34375 14.0511 8.34375 14.25V15.375C8.34375 15.5739 8.42277 15.7647 8.56342 15.9053C8.70407 16.046 8.89484 16.125 9.09375 16.125H10.5938C10.7927 16.125 10.9834 16.046 11.1241 15.9053C11.2647 15.7647 11.3438 15.5739 11.3438 15.375V14.25ZM9.09375 15.375V14.25H10.5938V15.375H9.09375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M6.84375 14.25C6.84375 14.0511 6.76473 13.8603 6.62408 13.7197C6.48343 13.579 6.29266 13.5 6.09375 13.5H4.59375C4.39484 13.5 4.20407 13.579 4.06342 13.7197C3.92277 13.8603 3.84375 14.0511 3.84375 14.25V15.375C3.84375 15.5739 3.92277 15.7647 4.06342 15.9053C4.20407 16.046 4.39484 16.125 4.59375 16.125H6.09375C6.29266 16.125 6.48343 16.046 6.62408 15.9053C6.76473 15.7647 6.84375 15.5739 6.84375 15.375V14.25ZM4.59375 15.375V14.25H6.09375V15.375H4.59375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M10.5938 16.875H9.09375C8.89484 16.875 8.70407 16.954 8.56342 17.0947C8.42277 17.2353 8.34375 17.4261 8.34375 17.625V18.75C8.34375 18.9489 8.42277 19.1397 8.56342 19.2803C8.70407 19.421 8.89484 19.5 9.09375 19.5H10.5938C10.7927 19.5 10.9834 19.421 11.1241 19.2803C11.2647 19.1397 11.3438 18.9489 11.3438 18.75V17.625C11.3438 17.4261 11.2647 17.2353 11.1241 17.0947C10.9834 16.954 10.7927 16.875 10.5938 16.875ZM9.09375 18.75V17.625H10.5938V18.75H9.09375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M13.5938 12.75H15.0938C15.2927 12.75 15.4834 12.671 15.6241 12.5303C15.7647 12.3897 15.8438 12.1989 15.8438 12V10.875C15.8438 10.6761 15.7647 10.4853 15.6241 10.3447C15.4834 10.204 15.2927 10.125 15.0938 10.125H13.5938C13.3948 10.125 13.2041 10.204 13.0634 10.3447C12.9228 10.4853 12.8438 10.6761 12.8438 10.875V12C12.8438 12.1989 12.9228 12.3897 13.0634 12.5303C13.2041 12.671 13.3948 12.75 13.5938 12.75ZM13.5938 10.875H15.0938V12H13.5938V10.875Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M12.8438 15.375C12.8438 15.5739 12.9228 15.7647 13.0634 15.9053C13.2041 16.046 13.3948 16.125 13.5938 16.125C13.6932 16.125 13.7886 16.0855 13.8589 16.0152C13.9292 15.9448 13.9688 15.8495 13.9688 15.75C13.9688 15.6505 13.9292 15.5552 13.8589 15.4848C13.7886 15.4145 13.6932 15.375 13.5938 15.375V14.25H15.0938C15.1932 14.25 15.2886 14.2105 15.3589 14.1402C15.4292 14.0698 15.4688 13.9745 15.4688 13.875C15.4688 13.7755 15.4292 13.6802 15.3589 13.6098C15.2886 13.5395 15.1932 13.5 15.0938 13.5H13.5938C13.3948 13.5 13.2041 13.579 13.0634 13.7197C12.9228 13.8603 12.8438 14.0511 12.8438 14.25V15.375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M6.09375 16.875H4.59375C4.39484 16.875 4.20407 16.954 4.06342 17.0947C3.92277 17.2353 3.84375 17.4261 3.84375 17.625V18.75C3.84375 18.9489 3.92277 19.1397 4.06342 19.2803C4.20407 19.421 4.39484 19.5 4.59375 19.5H6.09375C6.29266 19.5 6.48343 19.421 6.62408 19.2803C6.76473 19.1397 6.84375 18.9489 6.84375 18.75V17.625C6.84375 17.4261 6.76473 17.2353 6.62408 17.0947C6.48343 16.954 6.29266 16.875 6.09375 16.875ZM4.59375 18.75V17.625H6.09375V18.75H4.59375Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M18.4687 13.5C17.5046 13.5 16.562 13.7859 15.7603 14.3216C14.9587 14.8573 14.3338 15.6186 13.9648 16.5094C13.5959 17.4002 13.4993 18.3804 13.6874 19.3261C13.8755 20.2717 14.3398 21.1404 15.0216 21.8221C15.7034 22.5039 16.572 22.9682 17.5177 23.1563C18.4633 23.3444 19.4435 23.2479 20.3343 22.8789C21.2251 22.5099 21.9865 21.8851 22.5222 21.0834C23.0578 20.2817 23.3437 19.3392 23.3437 18.375C23.3422 17.0825 22.8282 15.8434 21.9142 14.9295C21.0003 14.0156 19.7612 13.5015 18.4687 13.5ZM18.4687 22.5C17.6529 22.5 16.8554 22.2581 16.177 21.8048C15.4987 21.3516 14.97 20.7073 14.6577 19.9536C14.3455 19.1998 14.2638 18.3704 14.423 17.5703C14.5822 16.7701 14.975 16.0351 15.5519 15.4582C16.1288 14.8813 16.8638 14.4884 17.664 14.3293C18.4642 14.1701 19.2936 14.2518 20.0473 14.564C20.8011 14.8762 21.4453 15.4049 21.8986 16.0833C22.3518 16.7616 22.5937 17.5592 22.5937 18.375C22.5924 19.4686 22.1574 20.5171 21.3841 21.2904C20.6108 22.0637 19.5624 22.4987 18.4687 22.5Z"
+                        fill="#FB5B45"
+                      />
+                      <path
+                        d="M20.6357 16.803L17.719 19.7198L16.3933 18.3938C16.3226 18.3254 16.2279 18.2876 16.1296 18.2885C16.0312 18.2894 15.9372 18.3288 15.8676 18.3983C15.7981 18.4679 15.7587 18.5619 15.7578 18.6602C15.757 18.7586 15.7948 18.8533 15.8631 18.924L17.4538 20.5151C17.5241 20.5854 17.6195 20.6249 17.719 20.6249C17.8184 20.6249 17.9138 20.5854 17.9841 20.5151L21.1659 17.3333C21.2343 17.2625 21.2721 17.1678 21.2712 17.0695C21.2703 16.9712 21.2309 16.8771 21.1614 16.8076C21.0919 16.738 20.9978 16.6986 20.8995 16.6978C20.8011 16.6969 20.7064 16.7347 20.6357 16.803Z"
+                        fill="#FB5B45"
+                      />
+                    </g>
+                    <defs>
+                      <clipPath id="clip0">
+                        <rect
+                          width="24"
+                          height="24"
+                          fill="white"
+                          transform="translate(0.09375)"
+                        />
+                      </clipPath>
+                    </defs>
+                  </svg>
+                </div>
+              </div>
+              <div class="card-body">
+                <div class="d-flex mb-3">
+                  <div>
+                    <img
+                      class="rounded-img-sm mr-2"
+                      src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                    />
+                  </div>
+                  <div class="d-block">
+                    <p class="m-0 p-0 title_1">Aboki Joke</p>
+                    <p class="m-0 p-0 title_2">Obstetrics & Gynaecology</p>
+                  </div>
+                </div>
+                <div class="d-flex justify-content-between title_2 py-4">
+                  <div>
+                    <p class="mb-1">Date:</p>
+                    <p class="m-0 pr-2">Tuesday 1 September, 2020</p>
+                  </div>
+                  <div>
+                    <p class="mb-1">Time:</p>
+                    <p class="m-0">8:00 AM - 9:00 AM</p>
+                  </div>
+                </div>
+              </div>
+              <div class="card-footer text-muted">
+                <div class="d-flex justify-content-between">
+                  <div class="title">Cancel Appointment</div>
+                  <div class="title_1">Reschedule</div>
+                </div>
+              </div>
+            </div>
+          </div>
+          <div class="col-md-8">
+            <div class="card">
+              <div class="card-header">
+                <div class="d-flex justify-content-between">
+                  <div>Recent Activities</div>
+                  <div>
+                    <svg
+                      width="25"
+                      height="24"
+                      viewBox="0 0 25 24"
+                      fill="none"
+                      xmlns="http://www.w3.org/2000/svg"
+                    >
+                      <g clip-path="url(#clip0)">
+                        <path
+                          d="M14.5419 1.5H17.7163V0.75H14.3413C14.2796 0.749991 14.2188 0.765217 14.1644 0.794327C14.11 0.823437 14.0636 0.865531 14.0293 0.916875L13.2793 2.04188L13.9033 2.45812L14.5419 1.5Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M11.1402 22.5H7.96582V23.25H11.3408C11.4025 23.25 11.4633 23.2348 11.5177 23.2057C11.5722 23.1766 11.6186 23.1345 11.6528 23.0831L12.4028 21.9581L11.7788 21.5419L11.1402 22.5Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M16.9658 9C15.6609 8.99832 14.3812 9.35924 13.2694 10.0425L13.0748 9.76987C13.1569 9.59805 13.2296 9.42192 13.2927 9.24225L14.7777 8.99475C14.8652 8.98011 14.9447 8.93493 15.002 8.86724C15.0593 8.79954 15.0908 8.71371 15.0908 8.625V6.375C15.0908 6.28629 15.0593 6.20046 15.002 6.13276C14.9447 6.06507 14.8652 6.01989 14.7777 6.00525L13.2927 5.75775C13.2296 5.57808 13.1569 5.40195 13.0748 5.23012L13.9493 4.00575C14.0009 3.93351 14.0251 3.84533 14.0178 3.7569C14.0104 3.66846 13.972 3.58549 13.9092 3.52275L12.3181 1.93163C12.2553 1.86887 12.1724 1.8304 12.0839 1.82305C11.9955 1.8157 11.9073 1.83996 11.8351 1.8915L10.6107 2.766C10.4388 2.68456 10.2627 2.61246 10.0831 2.55L9.83557 1.065C9.82133 0.977163 9.77632 0.89725 9.70859 0.839544C9.64085 0.781838 9.5548 0.750099 9.46582 0.75H7.21582C7.12711 0.750026 7.04128 0.7815 6.97358 0.838829C6.90589 0.896159 6.86071 0.975631 6.84607 1.06313L6.59857 2.54813C6.4189 2.6112 6.24277 2.68393 6.07095 2.766L4.84657 1.8915C4.77433 1.83996 4.68615 1.8157 4.59772 1.82305C4.50928 1.8304 4.42631 1.86887 4.36357 1.93163L2.77245 3.52275C2.70969 3.58549 2.67122 3.66846 2.66387 3.7569C2.65652 3.84533 2.68078 3.93351 2.73232 4.00575L3.61095 5.23575C3.52953 5.40517 3.45779 5.57908 3.39607 5.75663L1.90395 6.00525C1.81645 6.01989 1.73698 6.06507 1.67965 6.13276C1.62232 6.20046 1.59085 6.28629 1.59082 6.375V8.625C1.59085 8.71371 1.62232 8.79954 1.67965 8.86724C1.73698 8.93493 1.81645 8.98011 1.90395 8.99475L3.3957 9.24337C3.45713 9.42184 3.52759 9.59707 3.60682 9.76838L2.73232 10.9943C2.68078 11.0665 2.65652 11.1547 2.66387 11.2431C2.67122 11.3315 2.70969 11.4145 2.77245 11.4773L4.36357 13.0684C4.42631 13.1311 4.50928 13.1696 4.59772 13.177C4.68615 13.1843 4.77433 13.16 4.84657 13.1085L6.07095 12.234C6.24277 12.3161 6.4189 12.3888 6.59857 12.4519L6.84607 13.9369C6.86071 14.0244 6.90589 14.1038 6.97358 14.1612C7.04128 14.2185 7.12711 14.25 7.21582 14.25H9.46582C9.55453 14.25 9.64036 14.2185 9.70806 14.1612C9.77575 14.1038 9.82093 14.0244 9.83557 13.9369L10.0831 12.4519C10.2627 12.3888 10.4389 12.3161 10.6107 12.234L10.8833 12.4286C10.2001 13.5404 9.83914 14.8201 9.84082 16.125C9.84082 17.5342 10.2587 18.9117 11.0416 20.0834C11.8245 21.2551 12.9373 22.1684 14.2392 22.7076C15.5411 23.2469 16.9737 23.388 18.3558 23.1131C19.738 22.8382 21.0075 22.1596 22.004 21.1631C23.0004 20.1667 23.679 18.8971 23.9539 17.515C24.2288 16.1329 24.0877 14.7003 23.5485 13.3984C23.0092 12.0965 22.096 10.9837 20.9243 10.2008C19.7526 9.41787 18.375 9 16.9658 9ZM10.4663 11.4664C10.2033 11.607 9.92695 11.7211 9.64132 11.8069C9.57414 11.8269 9.51393 11.8654 9.46753 11.918C9.42113 11.9706 9.3904 12.0351 9.37882 12.1042L9.1482 13.5H7.53345L7.30095 12.1042C7.28937 12.0351 7.25863 11.9706 7.21223 11.918C7.16583 11.8654 7.10563 11.8269 7.03845 11.8069C6.75281 11.7211 6.47645 11.607 6.21345 11.4664C6.15178 11.4333 6.08208 11.4181 6.01223 11.4225C5.94238 11.4269 5.87517 11.4508 5.8182 11.4915L4.66695 12.3135L3.52732 11.172L4.34932 10.0208C4.38998 9.96378 4.41388 9.89656 4.41832 9.82671C4.42276 9.75686 4.40756 9.68716 4.37445 9.6255C4.23405 9.35583 4.1182 9.07407 4.02832 8.78362C4.00586 8.72073 3.96701 8.66498 3.91578 8.62212C3.86456 8.57927 3.80282 8.55088 3.73695 8.53988L2.34082 8.30737V6.69263L3.73657 6.46012C3.80244 6.44912 3.86418 6.42073 3.91541 6.37788C3.96663 6.33502 4.00548 6.27927 4.02795 6.21637C4.12451 5.92171 4.24322 5.63476 4.38307 5.358C4.41172 5.29759 4.4237 5.23062 4.41776 5.16403C4.41183 5.09744 4.3882 5.03364 4.34932 4.97925L3.52732 3.828L4.66882 2.6865L5.82007 3.5085C5.87704 3.54916 5.94426 3.57306 6.01411 3.5775C6.08396 3.58194 6.15366 3.56674 6.21532 3.53363C6.47832 3.39299 6.75469 3.27893 7.04032 3.19313C7.1075 3.17305 7.16771 3.13455 7.21411 3.08199C7.26051 3.02942 7.29124 2.9649 7.30282 2.89575L7.53345 1.5H9.1482L9.3807 2.89575C9.39227 2.9649 9.42301 3.02942 9.46941 3.08199C9.51581 3.13455 9.57602 3.17305 9.6432 3.19313C9.92883 3.27893 10.2052 3.39299 10.4682 3.53363C10.5299 3.56674 10.5996 3.58194 10.6694 3.5775C10.7393 3.57306 10.8065 3.54916 10.8634 3.5085L12.0147 2.6865L13.1562 3.828L12.3342 4.97925C12.2935 5.03622 12.2696 5.10344 12.2652 5.17329C12.2608 5.24314 12.276 5.31284 12.3091 5.3745C12.4497 5.6375 12.5638 5.91387 12.6496 6.1995C12.6696 6.26668 12.7081 6.32689 12.7607 6.37329C12.8133 6.41969 12.8778 6.45042 12.9469 6.462L14.3427 6.6945V8.30925L12.9469 8.54175C12.8778 8.55333 12.8133 8.58406 12.7607 8.63046C12.7081 8.67686 12.6696 8.73707 12.6496 8.80425C12.5638 9.08988 12.4497 9.36625 12.3091 9.62925C12.276 9.69091 12.2608 9.76061 12.2652 9.83046C12.2696 9.90031 12.2935 9.96753 12.3342 10.0245L13.1562 11.1757L12.0147 12.3173L10.8634 11.4952C10.8066 11.4536 10.7393 11.4289 10.669 11.4238C10.5988 11.4186 10.5285 11.4334 10.4663 11.4664ZM17.3408 22.4809V21.375H16.5908V22.4809C15.035 22.3878 13.5672 21.7278 12.4651 20.6257C11.363 19.5236 10.703 18.0558 10.6099 16.5H11.7158V15.75H10.6047C10.6622 14.7304 10.9676 13.7401 11.4942 12.8651L11.8351 13.1085C11.9073 13.16 11.9955 13.1843 12.0839 13.177C12.1724 13.1696 12.2553 13.1311 12.3181 13.0684L13.9092 11.4773C13.972 11.4145 14.0104 11.3315 14.0178 11.2431C14.0251 11.1547 14.0009 11.0665 13.9493 10.9943L13.7059 10.6534C14.5809 10.1267 15.5712 9.82141 16.5908 9.76388V10.875H17.3408V9.76912C18.8966 9.86219 20.3644 10.5222 21.4665 11.6243C22.5686 12.7264 23.2286 14.1942 23.3217 15.75H22.2158V16.5H23.3217C23.2286 18.0558 22.5686 19.5236 21.4665 20.6257C20.3644 21.7278 18.8966 22.3878 17.3408 22.4809Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M17.3408 11.625H16.5908V16.125C16.5908 16.2245 16.6303 16.3198 16.7007 16.3902C16.771 16.4605 16.8664 16.5 16.9658 16.5H19.9658V15.75H17.3408V11.625Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M10.2157 6.375H8.71572V4.125C8.71566 4.03938 8.68632 3.95636 8.63255 3.88973C8.57878 3.8231 8.50383 3.77688 8.42015 3.75874C8.33648 3.7406 8.24912 3.75165 8.17259 3.79003C8.09606 3.82842 8.03496 3.89184 7.99947 3.96975L6.12447 8.09475C6.09849 8.15187 6.08726 8.21458 6.09181 8.27716C6.09635 8.33974 6.11652 8.40018 6.15048 8.45294C6.18444 8.5057 6.23109 8.5491 6.28616 8.57916C6.34124 8.60922 6.40297 8.62498 6.46572 8.625H7.96572V10.875C7.96565 10.9607 7.99494 11.0438 8.04871 11.1106C8.10247 11.1773 8.17747 11.2236 8.26122 11.2418C8.28736 11.2473 8.314 11.25 8.34072 11.25C8.41269 11.25 8.48313 11.2293 8.54364 11.1903C8.60415 11.1513 8.65217 11.0958 8.68197 11.0303L10.557 6.90525C10.5829 6.84814 10.5942 6.78542 10.5896 6.72284C10.5851 6.66026 10.5649 6.59982 10.531 6.54706C10.497 6.4943 10.4503 6.4509 10.3953 6.42084C10.3402 6.39078 10.2785 6.37502 10.2157 6.375ZM8.71572 9.14363V8.25C8.71572 8.15055 8.67621 8.05516 8.60588 7.98484C8.53556 7.91451 8.44017 7.875 8.34072 7.875H7.04809L7.96572 5.85638V6.75C7.96572 6.84946 8.00522 6.94484 8.07555 7.01517C8.14588 7.08549 8.24126 7.125 8.34072 7.125H9.63334L8.71572 9.14363Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M8.71582 2.25H7.96582V3H8.71582V2.25Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M4.89361 3.52267L4.36328 4.053L4.89361 4.58333L5.42394 4.053L4.89361 3.52267Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M3.84082 7.125H3.09082V7.875H3.84082V7.125Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M4.89361 10.4165L4.36328 10.9468L4.89361 11.4771L5.42394 10.9468L4.89361 10.4165Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M8.71582 12H7.96582V12.75H8.71582V12Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M11.7881 10.4167L11.2578 10.947L11.7881 11.4773L12.3185 10.947L11.7881 10.4167Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M13.5908 7.125H12.8408V7.875H13.5908V7.125Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M11.7881 3.52288L11.2578 4.05321L11.7881 4.58354L12.3185 4.05321L11.7881 3.52288Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M20.7158 2.625C20.4192 2.625 20.1291 2.71297 19.8825 2.8778C19.6358 3.04262 19.4435 3.27689 19.33 3.55097C19.2165 3.82506 19.1868 4.12666 19.2446 4.41764C19.3025 4.70861 19.4454 4.97588 19.6552 5.18566C19.8649 5.39544 20.1322 5.5383 20.4232 5.59618C20.7142 5.65406 21.0158 5.62435 21.2899 5.51082C21.5639 5.39729 21.7982 5.20503 21.963 4.95836C22.1279 4.71168 22.2158 4.42167 22.2158 4.125C22.2158 3.72718 22.0578 3.34564 21.7765 3.06434C21.4952 2.78304 21.1137 2.625 20.7158 2.625ZM20.7158 4.875C20.5675 4.875 20.4225 4.83101 20.2991 4.7486C20.1758 4.66619 20.0797 4.54906 20.0229 4.41201C19.9661 4.27497 19.9513 4.12417 19.9802 3.97868C20.0092 3.8332 20.0806 3.69956 20.1855 3.59467C20.2904 3.48978 20.424 3.41835 20.5695 3.38941C20.715 3.36047 20.8658 3.37532 21.0028 3.43209C21.1399 3.48886 21.257 3.58499 21.3394 3.70832C21.4218 3.83166 21.4658 3.97666 21.4658 4.125C21.4658 4.32391 21.3868 4.51468 21.2462 4.65533C21.1055 4.79598 20.9147 4.875 20.7158 4.875Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M23.8069 3.01125L23.5561 2.94863L23.6892 2.727C23.7322 2.65525 23.75 2.57116 23.7398 2.48813C23.7295 2.40509 23.6918 2.32787 23.6326 2.26875L22.5721 1.20825C22.513 1.14904 22.4357 1.11131 22.3527 1.10105C22.2697 1.09079 22.1856 1.10859 22.1138 1.15163L21.8922 1.28475L21.8296 1.03387C21.8093 0.952783 21.7624 0.880803 21.6965 0.829371C21.6306 0.777938 21.5494 0.750002 21.4658 0.75H19.9658C19.8822 0.750002 19.801 0.777938 19.7351 0.829371C19.6692 0.880803 19.6224 0.952783 19.6021 1.03387L19.5394 1.28475L19.3178 1.15163C19.2461 1.10859 19.162 1.09079 19.0789 1.10105C18.9959 1.11131 18.9187 1.14904 18.8596 1.20825L17.7991 2.26875C17.7399 2.32787 17.7021 2.40509 17.6919 2.48813C17.6816 2.57116 17.6994 2.65525 17.7424 2.727L17.8756 2.94863L17.6247 3.01125C17.5436 3.03156 17.4716 3.07839 17.4202 3.1443C17.3688 3.2102 17.3408 3.2914 17.3408 3.375V4.875C17.3408 4.9586 17.3688 5.0398 17.4202 5.1057C17.4716 5.17161 17.5436 5.21844 17.6247 5.23875L17.8756 5.30137L17.7424 5.523C17.6994 5.59475 17.6816 5.67884 17.6919 5.76187C17.7021 5.84491 17.7399 5.92213 17.7991 5.98125L18.8596 7.04175C18.9187 7.10096 18.9959 7.13869 19.0789 7.14895C19.162 7.15921 19.2461 7.14141 19.3178 7.09838L19.5394 6.96525L19.6021 7.21613C19.6224 7.29722 19.6692 7.3692 19.7351 7.42063C19.801 7.47206 19.8822 7.5 19.9658 7.5H21.4658C21.5494 7.5 21.6306 7.47206 21.6965 7.42063C21.7624 7.3692 21.8093 7.29722 21.8296 7.21613L21.8922 6.96525L22.1138 7.09838C22.1856 7.14141 22.2697 7.15921 22.3527 7.14895C22.4357 7.13869 22.513 7.10096 22.5721 7.04175L23.6326 5.98125C23.6918 5.92213 23.7295 5.84491 23.7398 5.76187C23.75 5.67884 23.7322 5.59475 23.6892 5.523L23.5561 5.30137L23.8069 5.23875C23.888 5.21844 23.96 5.17161 24.0114 5.1057C24.0629 5.0398 24.0908 4.9586 24.0908 4.875V3.375C24.0908 3.2914 24.0629 3.2102 24.0114 3.1443C23.96 3.07839 23.888 3.03156 23.8069 3.01125ZM23.3408 4.58212L22.8829 4.69462C22.8261 4.70887 22.7734 4.73623 22.7291 4.77452C22.6847 4.8128 22.65 4.86094 22.6276 4.91508C22.6052 4.96922 22.5958 5.02785 22.6002 5.08627C22.6045 5.14469 22.6225 5.20128 22.6527 5.2515L22.8953 5.65613L22.2488 6.30262L21.8442 6.06C21.794 6.02983 21.7374 6.01184 21.679 6.00748C21.6205 6.00312 21.5619 6.01251 21.5078 6.0349C21.4536 6.05728 21.4055 6.09204 21.3672 6.13638C21.3289 6.18073 21.3016 6.23342 21.2873 6.29025L21.1729 6.75H20.2587L20.1462 6.29212C20.1319 6.2353 20.1046 6.1826 20.0663 6.13826C20.028 6.09391 19.9799 6.05916 19.9257 6.03677C19.8716 6.01439 19.813 6.005 19.7545 6.00936C19.6961 6.01372 19.6395 6.0317 19.5893 6.06188L19.1847 6.3045L18.5382 5.658L18.7808 5.25338C18.811 5.20316 18.829 5.14657 18.8333 5.08815C18.8377 5.02973 18.8283 4.9711 18.8059 4.91696C18.7835 4.86282 18.7488 4.81468 18.7044 4.77639C18.6601 4.73811 18.6074 4.71075 18.5506 4.6965L18.0908 4.58212V3.66788L18.5487 3.55538C18.6055 3.54113 18.6582 3.51377 18.7026 3.47548C18.7469 3.4372 18.7817 3.38906 18.804 3.33492C18.8264 3.28078 18.8358 3.22215 18.8315 3.16373C18.8271 3.1053 18.8091 3.04872 18.7789 2.9985L18.5363 2.59387L19.1828 1.94737L19.5874 2.19C19.6377 2.22017 19.6943 2.23814 19.7527 2.24248C19.8112 2.24682 19.8698 2.2374 19.924 2.21497C19.9781 2.19255 20.0263 2.15775 20.0645 2.11336C20.1028 2.06897 20.1301 2.01623 20.1443 1.95937L20.2568 1.50188H21.1711L21.2836 1.95937C21.2978 2.01623 21.3251 2.06897 21.3634 2.11336C21.4016 2.15775 21.4498 2.19255 21.5039 2.21497C21.5581 2.2374 21.6167 2.24682 21.6751 2.24248C21.7336 2.23814 21.7902 2.22017 21.8404 2.19L22.2451 1.94737L22.8916 2.59387L22.6489 2.9985C22.6188 3.04872 22.6008 3.1053 22.5964 3.16373C22.5921 3.22215 22.6015 3.28078 22.6238 3.33492C22.6462 3.38906 22.681 3.4372 22.7253 3.47548C22.7697 3.51377 22.8224 3.54113 22.8792 3.55538L23.3371 3.66788L23.3408 4.58212Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M4.96582 18.375C4.66915 18.375 4.37914 18.463 4.13247 18.6278C3.88579 18.7926 3.69353 19.0269 3.58 19.301C3.46647 19.5751 3.43677 19.8767 3.49464 20.1676C3.55252 20.4586 3.69538 20.7259 3.90516 20.9357C4.11494 21.1454 4.38221 21.2883 4.67319 21.3462C4.96416 21.4041 5.26576 21.3744 5.53985 21.2608C5.81394 21.1473 6.0482 20.955 6.21303 20.7084C6.37785 20.4617 6.46582 20.1717 6.46582 19.875C6.46582 19.4772 6.30779 19.0956 6.02648 18.8143C5.74518 18.533 5.36365 18.375 4.96582 18.375ZM4.96582 20.625C4.81749 20.625 4.67248 20.581 4.54914 20.4986C4.42581 20.4162 4.32968 20.2991 4.27291 20.162C4.21615 20.025 4.20129 19.8742 4.23023 19.7287C4.25917 19.5832 4.3306 19.4496 4.43549 19.3447C4.54038 19.2398 4.67402 19.1683 4.8195 19.1394C4.96499 19.1105 5.11579 19.1253 5.25283 19.1821C5.38988 19.2389 5.50701 19.335 5.58942 19.4583C5.67183 19.5817 5.71582 19.7267 5.71582 19.875C5.71582 20.0739 5.6368 20.2647 5.49615 20.4053C5.3555 20.546 5.16473 20.625 4.96582 20.625Z"
+                          fill="#FB5B45"
+                        />
+                        <path
+                          d="M7.9392 21.273L7.80607 21.0514L8.05695 20.9888C8.13804 20.9684 8.21002 20.9216 8.26145 20.8557C8.31288 20.7898 8.34082 20.7086 8.34082 20.625V19.125C8.34082 19.0414 8.31288 18.9602 8.26145 18.8943C8.21002 18.8284 8.13804 18.7816 8.05695 18.7612L7.80607 18.6986L7.9392 18.477C7.98223 18.4052 8.00003 18.3212 7.98977 18.2381C7.97951 18.1551 7.94178 18.0779 7.88257 18.0187L6.82207 16.9583C6.76295 16.899 6.68573 16.8613 6.60269 16.851C6.51966 16.8408 6.43557 16.8586 6.36382 16.9016L6.1422 17.0347L6.07957 16.7839C6.05926 16.7028 6.01243 16.6308 5.94652 16.5794C5.88062 16.5279 5.79942 16.5 5.71582 16.5H4.21582C4.13222 16.5 4.05102 16.5279 3.98512 16.5794C3.91921 16.6308 3.87238 16.7028 3.85207 16.7839L3.78945 17.0347L3.56782 16.9016C3.49607 16.8586 3.41198 16.8408 3.32895 16.851C3.24591 16.8613 3.16869 16.899 3.10957 16.9583L2.04907 18.0187C1.98986 18.0779 1.95213 18.1551 1.94187 18.2381C1.93161 18.3212 1.94941 18.4052 1.99245 18.477L2.12557 18.6986L1.8747 18.7612C1.7936 18.7816 1.72162 18.8284 1.67019 18.8943C1.61876 18.9602 1.59082 19.0414 1.59082 19.125V20.625C1.59082 20.7086 1.61876 20.7898 1.67019 20.8557C1.72162 20.9216 1.7936 20.9684 1.8747 20.9888L2.12557 21.0514L1.99245 21.273C1.94941 21.3448 1.93161 21.4288 1.94187 21.5119C1.95213 21.5949 1.98986 21.6721 2.04907 21.7313L3.10957 22.7917C3.16869 22.851 3.24591 22.8887 3.32895 22.899C3.41198 22.9092 3.49607 22.8914 3.56782 22.8484L3.78945 22.7153L3.85207 22.9661C3.87238 23.0472 3.91921 23.1192 3.98512 23.1706C4.05102 23.2221 4.13222 23.25 4.21582 23.25H5.71582C5.79942 23.25 5.88062 23.2221 5.94652 23.1706C6.01243 23.1192 6.05926 23.0472 6.07957 22.9661L6.1422 22.7153L6.36382 22.8484C6.43557 22.8914 6.51966 22.9092 6.60269 22.899C6.68573 22.8887 6.76295 22.851 6.82207 22.7917L7.88257 21.7313C7.94178 21.6721 7.97951 21.5949 7.98977 21.5119C8.00003 21.4288 7.98223 21.3448 7.9392 21.273ZM7.59082 20.3321L7.13295 20.4446C7.07612 20.4589 7.02342 20.4862 6.97908 20.5245C6.93473 20.5628 6.89998 20.6109 6.87759 20.6651C6.85521 20.7192 6.84582 20.7779 6.85018 20.8363C6.85454 20.8947 6.87252 20.9513 6.9027 21.0015L7.14532 21.4061L6.49882 22.0526L6.0942 21.81C6.04398 21.7798 5.98739 21.7618 5.92897 21.7575C5.87055 21.7531 5.81192 21.7625 5.75778 21.7849C5.70364 21.8073 5.6555 21.842 5.61721 21.8864C5.57893 21.9307 5.55157 21.9834 5.53732 22.0402L5.42295 22.5H4.5087L4.3962 22.0421C4.38195 21.9853 4.35459 21.9326 4.3163 21.8883C4.27802 21.8439 4.22988 21.8092 4.17574 21.7868C4.1216 21.7644 4.06297 21.755 4.00455 21.7594C3.94613 21.7637 3.88954 21.7817 3.83932 21.8119L3.4347 22.0545L2.78632 21.408L3.02895 21.0034C3.05912 20.9532 3.0771 20.8966 3.08146 20.8381C3.08582 20.7797 3.07643 20.7211 3.05405 20.667C3.03166 20.6128 2.99691 20.5647 2.95256 20.5264C2.90822 20.4881 2.85552 20.4607 2.7987 20.4465L2.34082 20.3321V19.4179L2.7987 19.3054C2.85552 19.2911 2.90822 19.2638 2.95256 19.2255C2.99691 19.1872 3.03166 19.1391 3.05405 19.0849C3.07643 19.0308 3.08582 18.9721 3.08146 18.9137C3.0771 18.8553 3.05912 18.7987 3.02895 18.7485L2.78632 18.3439L3.43282 17.6974L3.83745 17.94C3.88769 17.9702 3.9443 17.9881 4.00275 17.9925C4.06119 17.9968 4.11984 17.9874 4.17398 17.965C4.22813 17.9425 4.27626 17.9077 4.31453 17.8634C4.35279 17.819 4.38012 17.7662 4.39432 17.7094L4.50682 17.2519H5.42107L5.53357 17.7094C5.54777 17.7662 5.5751 17.819 5.61336 17.8634C5.65163 17.9077 5.69976 17.9425 5.75391 17.965C5.80805 17.9874 5.8667 17.9968 5.92515 17.9925C5.98359 17.9881 6.0402 17.9702 6.09045 17.94L6.49507 17.6974L7.14157 18.3439L6.89895 18.7485C6.86877 18.7987 6.85079 18.8553 6.84643 18.9137C6.84207 18.9721 6.85146 19.0308 6.87384 19.0849C6.89623 19.1391 6.93098 19.1872 6.97533 19.2255C7.01967 19.2638 7.07237 19.2911 7.1292 19.3054L7.58707 19.4179L7.59082 20.3321Z"
+                          fill="#FB5B45"
+                        />
+                      </g>
+                      <defs>
+                        <clipPath id="clip0">
+                          <rect
+                            width="24"
+                            height="24"
+                            fill="white"
+                            transform="translate(0.84082)"
+                          />
+                        </clipPath>
+                      </defs>
+                    </svg>
+                  </div>
+                </div>
+              </div>
+              <div class="card-body">
+                <div class="row mb-2">
+                  <div class="col-md-3">
+                    <div class="d-flex mb-3">
+                      <div>
+                        <img
+                          class="rounded-img-sm mr-2"
+                          src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                        />
+                      </div>
+                      <div class="">
+                        <p class="m-0 p-0 title_1">Aboki Joke</p>
+                        <p class="m-0 p-0 title_2">03 Aug, 2020</p>
+                      </div>
+                    </div>
+                  </div>
+                  <div class="col-md-9">
+                    <div class="">
+                      <div class="m-0 title_2">
+                        Dr Elliot updated your medical records on Friday 4th
+                        August, 2020 based on your complaints and diagnosis.
+                      </div>
+                    </div>
+                  </div>
+                </div>
+                <div class="row">
+                  <div class="col-md-3">
+                    <div class="d-flex mb-3">
+                      <div>
+                        <img
+                          class="rounded-img-sm mr-2"
+                          src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                        />
+                      </div>
+                      <div class="">
+                        <p class="m-0 p-0 title_1">Aboki Joke</p>
+                        <p class="m-0 p-0 title_2">03 Aug, 2020</p>
+                      </div>
+                    </div>
+                  </div>
+                  <div class="col-md-9">
+                    <div class="">
+                      <div class="m-0 title_2">
+                        Dr Elliot updated your medical records on Friday 4th
+                        August, 2020 based on your complaints and diagnosis.
+                      </div>
+                    </div>
+                  </div>
+                </div>
+              </div>
+              <div class="card-footer text-muted">
+                <div class="title text-right">See All Activities</div>
+              </div>
+            </div>
+          </div>
+        </div>
+
+        
+
+        <div class="row pb-4">
+              <div class="col-md-4 col-sm-12">
+                <div class="heading-title my-4 pl-4">Current Medication (4)</div>
+                <div className="">
+                  <div class="card">
+                    <ul class="list-group list-group-flush">
+                      <li class="list-group-item">
+                        <div class="d-flex">
+                          <div class="mr-2">
+                            <svg
+                              width="25"
+                              height="25"
+                              viewBox="0 0 25 25"
+                              fill="none"
+                              xmlns="http://www.w3.org/2000/svg"
+                            >
+                              <g clip-path="url(#clip0)">
+                                <path
+                                  d="M23.1941 10.2749L15.1781 2.25913C14.2903 1.37761 13.0893 0.883886 11.8383 0.886101C10.5872 0.888315 9.38795 1.38629 8.50329 2.27095C7.61864 3.1556 7.12066 4.35482 7.11845 5.60591C7.11623 6.857 7.60996 8.05798 8.49148 8.94576L16.5075 16.9625C17.3942 17.8492 18.5968 18.3473 19.8508 18.3473C21.1048 18.3473 22.3075 17.8491 23.1942 16.9624C24.0809 16.0757 24.579 14.873 24.579 13.619C24.579 12.365 24.0808 11.1624 23.1941 10.2757V10.2749ZM7.76285 5.60245C7.76351 4.79737 8.00272 4.01055 8.45026 3.34133C8.89781 2.67212 9.53362 2.15053 10.2774 1.84243C11.0212 1.53433 11.8396 1.45355 12.6293 1.61027C13.4189 1.767 14.1445 2.15421 14.7142 2.72301L18.4903 6.49907L12.7314 12.2578L8.95554 8.4817C8.57621 8.10449 8.27548 7.65578 8.07075 7.16155C7.86603 6.66733 7.76137 6.1374 7.76285 5.60245ZM22.7302 16.4977C21.9658 17.26 20.9304 17.6881 19.8509 17.6881C18.7714 17.6881 17.7359 17.26 16.9715 16.4977L13.1955 12.7218L18.954 6.96295L22.7302 10.7388C23.4926 11.5032 23.9207 12.5387 23.9207 13.6183C23.9207 14.6978 23.4926 15.7333 22.7302 16.4977Z"
+                                  fill="#FB5B45"
+                                />
+                                <path
+                                  d="M13.1488 4.28843C13.1793 4.3189 13.2155 4.34307 13.2553 4.35956C13.2951 4.37605 13.3378 4.38454 13.3809 4.38454C13.424 4.38454 13.4666 4.37605 13.5064 4.35956C13.5462 4.34307 13.5824 4.3189 13.6129 4.28843C13.6434 4.25796 13.6675 4.22178 13.684 4.18197C13.7005 4.14216 13.709 4.09949 13.709 4.0564C13.709 4.01331 13.7005 3.97064 13.684 3.93082C13.6675 3.89101 13.6434 3.85484 13.6129 3.82437C13.141 3.35355 12.5016 3.08914 11.835 3.08914C11.1684 3.08914 10.529 3.35355 10.0571 3.82437C10.0258 3.85463 10.0008 3.89083 9.98357 3.93086C9.96637 3.97089 9.95731 4.01394 9.95692 4.05751C9.95653 4.10108 9.96483 4.14429 9.98132 4.18462C9.99781 4.22495 10.0222 4.26159 10.053 4.2924C10.0838 4.32322 10.1204 4.34759 10.1607 4.36409C10.201 4.3806 10.2443 4.38891 10.2878 4.38854C10.3314 4.38817 10.3744 4.37913 10.4145 4.36194C10.4545 4.34475 10.4907 4.31976 10.521 4.28843C10.8698 3.94055 11.3423 3.74519 11.8349 3.74519C12.3275 3.74519 12.8 3.94055 13.1488 4.28843Z"
+                                  fill="#FB5B45"
+                                />
+                                <path
+                                  d="M6.12786 12.653C5.03039 12.653 3.95755 12.9784 3.04502 13.5881C2.13249 14.1978 1.42125 15.0644 1.00124 16.0783C0.581226 17.0922 0.471313 18.2079 0.685396 19.2843C0.89948 20.3607 1.42794 21.3494 2.20396 22.1255C2.97998 22.9015 3.96869 23.43 5.04507 23.6441C6.12146 23.8583 7.23716 23.7484 8.25109 23.3284C9.26503 22.9084 10.1317 22.1972 10.7414 21.2847C11.3511 20.3722 11.6765 19.2994 11.6765 18.2019C11.6749 16.7308 11.0898 15.3204 10.0496 14.2801C9.00936 13.2399 7.59898 12.6547 6.12786 12.653ZM6.12786 23.0945C5.16018 23.0946 4.21423 22.8076 3.40961 22.2701C2.605 21.7325 1.97787 20.9684 1.60753 20.0744C1.23719 19.1803 1.14027 18.1966 1.32904 17.2475C1.5178 16.2984 1.98376 15.4266 2.668 14.7424C3.35224 14.0581 4.22402 13.5921 5.1731 13.4033C6.12218 13.2145 7.10594 13.3114 7.99996 13.6817C8.89398 14.052 9.65811 14.6791 10.1957 15.4837C10.7333 16.2883 11.0203 17.2342 11.0203 18.2019C11.0186 19.4989 10.5026 20.7423 9.58544 21.6593C8.66828 22.5764 7.42485 23.0922 6.12786 23.0938V23.0945Z"
+                                  fill="#FB5B45"
+                                />
+                                <path
+                                  d="M8.1107 15.755L3.68101 20.1847C3.65054 20.2152 3.62637 20.2513 3.60988 20.2912C3.59339 20.331 3.5849 20.3736 3.5849 20.4167C3.5849 20.4598 3.59339 20.5025 3.60988 20.5423C3.62637 20.5821 3.65054 20.6183 3.68101 20.6488C3.71148 20.6792 3.74766 20.7034 3.78747 20.7199C3.82728 20.7364 3.86995 20.7449 3.91304 20.7449C3.95613 20.7449 3.9988 20.7364 4.03862 20.7199C4.07843 20.7034 4.1146 20.6792 4.14507 20.6488L8.57476 16.2191C8.6363 16.1575 8.67087 16.0741 8.67087 15.987C8.67087 15.9 8.6363 15.8165 8.57476 15.755C8.51322 15.6935 8.42976 15.6589 8.34273 15.6589C8.2557 15.6589 8.17224 15.6935 8.1107 15.755Z"
+                                  fill="#FB5B45"
+                                />
+                              </g>
+                              <defs>
+                                <clipPath id="clip0">
+                                  <rect
+                                    width="24"
+                                    height="24"
+                                    fill="white"
+                                    transform="translate(0.578979 0.3125)"
+                                  />
+                                </clipPath>
+                              </defs>
+                            </svg>
+                          </div>
+                          <div class="">
+                            <div class="title_1">Paracetamol</div>
+                            <div class="title_2">
+                              1 Pill (s); thrice Daily; 3 days
+                            </div>
+                          </div>
+                        </div>
+                      </li>
+                      <li class="list-group-item">
+                        <div class="d-flex">
+                          <div class="mr-2">
+                            <svg
+                              width="25"
+                              height="25"
+                              viewBox="0 0 25 25"
+                              fill="none"
+                              xmlns="http://www.w3.org/2000/svg"
+                            >
+                              <g clip-path="url(#clip0)">
+                                <path
+                                  d="M23.1941 10.2749L15.1781 2.25913C14.2903 1.37761 13.0893 0.883886 11.8383 0.886101C10.5872 0.888315 9.38795 1.38629 8.50329 2.27095C7.61864 3.1556 7.12066 4.35482 7.11845 5.60591C7.11623 6.857 7.60996 8.05798 8.49148 8.94576L16.5075 16.9625C17.3942 17.8492 18.5968 18.3473 19.8508 18.3473C21.1048 18.3473 22.3075 17.8491 23.1942 16.9624C24.0809 16.0757 24.579 14.873 24.579 13.619C24.579 12.365 24.0808 11.1624 23.1941 10.2757V10.2749ZM7.76285 5.60245C7.76351 4.79737 8.00272 4.01055 8.45026 3.34133C8.89781 2.67212 9.53362 2.15053 10.2774 1.84243C11.0212 1.53433 11.8396 1.45355 12.6293 1.61027C13.4189 1.767 14.1445 2.15421 14.7142 2.72301L18.4903 6.49907L12.7314 12.2578L8.95554 8.4817C8.57621 8.10449 8.27548 7.65578 8.07075 7.16155C7.86603 6.66733 7.76137 6.1374 7.76285 5.60245ZM22.7302 16.4977C21.9658 17.26 20.9304 17.6881 19.8509 17.6881C18.7714 17.6881 17.7359 17.26 16.9715 16.4977L13.1955 12.7218L18.954 6.96295L22.7302 10.7388C23.4926 11.5032 23.9207 12.5387 23.9207 13.6183C23.9207 14.6978 23.4926 15.7333 22.7302 16.4977Z"
+                                  fill="#FB5B45"
+                                />
+                                <path
+                                  d="M13.1488 4.28843C13.1793 4.3189 13.2155 4.34307 13.2553 4.35956C13.2951 4.37605 13.3378 4.38454 13.3809 4.38454C13.424 4.38454 13.4666 4.37605 13.5064 4.35956C13.5462 4.34307 13.5824 4.3189 13.6129 4.28843C13.6434 4.25796 13.6675 4.22178 13.684 4.18197C13.7005 4.14216 13.709 4.09949 13.709 4.0564C13.709 4.01331 13.7005 3.97064 13.684 3.93082C13.6675 3.89101 13.6434 3.85484 13.6129 3.82437C13.141 3.35355 12.5016 3.08914 11.835 3.08914C11.1684 3.08914 10.529 3.35355 10.0571 3.82437C10.0258 3.85463 10.0008 3.89083 9.98357 3.93086C9.96637 3.97089 9.95731 4.01394 9.95692 4.05751C9.95653 4.10108 9.96483 4.14429 9.98132 4.18462C9.99781 4.22495 10.0222 4.26159 10.053 4.2924C10.0838 4.32322 10.1204 4.34759 10.1607 4.36409C10.201 4.3806 10.2443 4.38891 10.2878 4.38854C10.3314 4.38817 10.3744 4.37913 10.4145 4.36194C10.4545 4.34475 10.4907 4.31976 10.521 4.28843C10.8698 3.94055 11.3423 3.74519 11.8349 3.74519C12.3275 3.74519 12.8 3.94055 13.1488 4.28843Z"
+                                  fill="#FB5B45"
+                                />
+                                <path
+                                  d="M6.12786 12.653C5.03039 12.653 3.95755 12.9784 3.04502 13.5881C2.13249 14.1978 1.42125 15.0644 1.00124 16.0783C0.581226 17.0922 0.471313 18.2079 0.685396 19.2843C0.89948 20.3607 1.42794 21.3494 2.20396 22.1255C2.97998 22.9015 3.96869 23.43 5.04507 23.6441C6.12146 23.8583 7.23716 23.7484 8.25109 23.3284C9.26503 22.9084 10.1317 22.1972 10.7414 21.2847C11.3511 20.3722 11.6765 19.2994 11.6765 18.2019C11.6749 16.7308 11.0898 15.3204 10.0496 14.2801C9.00936 13.2399 7.59898 12.6547 6.12786 12.653ZM6.12786 23.0945C5.16018 23.0946 4.21423 22.8076 3.40961 22.2701C2.605 21.7325 1.97787 20.9684 1.60753 20.0744C1.23719 19.1803 1.14027 18.1966 1.32904 17.2475C1.5178 16.2984 1.98376 15.4266 2.668 14.7424C3.35224 14.0581 4.22402 13.5921 5.1731 13.4033C6.12218 13.2145 7.10594 13.3114 7.99996 13.6817C8.89398 14.052 9.65811 14.6791 10.1957 15.4837C10.7333 16.2883 11.0203 17.2342 11.0203 18.2019C11.0186 19.4989 10.5026 20.7423 9.58544 21.6593C8.66828 22.5764 7.42485 23.0922 6.12786 23.0938V23.0945Z"
+                                  fill="#FB5B45"
+                                />
+                                <path
+                                  d="M8.1107 15.755L3.68101 20.1847C3.65054 20.2152 3.62637 20.2513 3.60988 20.2912C3.59339 20.331 3.5849 20.3736 3.5849 20.4167C3.5849 20.4598 3.59339 20.5025 3.60988 20.5423C3.62637 20.5821 3.65054 20.6183 3.68101 20.6488C3.71148 20.6792 3.74766 20.7034 3.78747 20.7199C3.82728 20.7364 3.86995 20.7449 3.91304 20.7449C3.95613 20.7449 3.9988 20.7364 4.03862 20.7199C4.07843 20.7034 4.1146 20.6792 4.14507 20.6488L8.57476 16.2191C8.6363 16.1575 8.67087 16.0741 8.67087 15.987C8.67087 15.9 8.6363 15.8165 8.57476 15.755C8.51322 15.6935 8.42976 15.6589 8.34273 15.6589C8.2557 15.6589 8.17224 15.6935 8.1107 15.755Z"
+                                  fill="#FB5B45"
+                                />
+                              </g>
+                              <defs>
+                                <clipPath id="clip0">
+                                  <rect
+                                    width="24"
+                                    height="24"
+                                    fill="white"
+                                    transform="translate(0.578979 0.3125)"
+                                  />
+                                </clipPath>
+                              </defs>
+                            </svg>
+                          </div>
+                          <div class="">
+                            <div class="title_1">Leonart Soft Gel (300 mg)</div>
+                            <div class="title_2">
+                              1 Pill (s); thrice Daily; 3 days
+                            </div>
+                          </div>
+                        </div>
+                      </li>
+                      <li class="list-group-item">
+                        <div class="d-flex">
+                          <div class="mr-2">
+                            <svg
+                              width="25"
+                              height="25"
+                              viewBox="0 0 25 25"
+                              fill="none"
+                              xmlns="http://www.w3.org/2000/svg"
+                            >
+                              <g clip-path="url(#clip0)">
+                                <path
+                                  d="M23.1941 10.2749L15.1781 2.25913C14.2903 1.37761 13.0893 0.883886 11.8383 0.886101C10.5872 0.888315 9.38795 1.38629 8.50329 2.27095C7.61864 3.1556 7.12066 4.35482 7.11845 5.60591C7.11623 6.857 7.60996 8.05798 8.49148 8.94576L16.5075 16.9625C17.3942 17.8492 18.5968 18.3473 19.8508 18.3473C21.1048 18.3473 22.3075 17.8491 23.1942 16.9624C24.0809 16.0757 24.579 14.873 24.579 13.619C24.579 12.365 24.0808 11.1624 23.1941 10.2757V10.2749ZM7.76285 5.60245C7.76351 4.79737 8.00272 4.01055 8.45026 3.34133C8.89781 2.67212 9.53362 2.15053 10.2774 1.84243C11.0212 1.53433 11.8396 1.45355 12.6293 1.61027C13.4189 1.767 14.1445 2.15421 14.7142 2.72301L18.4903 6.49907L12.7314 12.2578L8.95554 8.4817C8.57621 8.10449 8.27548 7.65578 8.07075 7.16155C7.86603 6.66733 7.76137 6.1374 7.76285 5.60245ZM22.7302 16.4977C21.9658 17.26 20.9304 17.6881 19.8509 17.6881C18.7714 17.6881 17.7359 17.26 16.9715 16.4977L13.1955 12.7218L18.954 6.96295L22.7302 10.7388C23.4926 11.5032 23.9207 12.5387 23.9207 13.6183C23.9207 14.6978 23.4926 15.7333 22.7302 16.4977Z"
+                                  fill="#FB5B45"
+                                />
+                                <path
+                                  d="M13.1488 4.28843C13.1793 4.3189 13.2155 4.34307 13.2553 4.35956C13.2951 4.37605 13.3378 4.38454 13.3809 4.38454C13.424 4.38454 13.4666 4.37605 13.5064 4.35956C13.5462 4.34307 13.5824 4.3189 13.6129 4.28843C13.6434 4.25796 13.6675 4.22178 13.684 4.18197C13.7005 4.14216 13.709 4.09949 13.709 4.0564C13.709 4.01331 13.7005 3.97064 13.684 3.93082C13.6675 3.89101 13.6434 3.85484 13.6129 3.82437C13.141 3.35355 12.5016 3.08914 11.835 3.08914C11.1684 3.08914 10.529 3.35355 10.0571 3.82437C10.0258 3.85463 10.0008 3.89083 9.98357 3.93086C9.96637 3.97089 9.95731 4.01394 9.95692 4.05751C9.95653 4.10108 9.96483 4.14429 9.98132 4.18462C9.99781 4.22495 10.0222 4.26159 10.053 4.2924C10.0838 4.32322 10.1204 4.34759 10.1607 4.36409C10.201 4.3806 10.2443 4.38891 10.2878 4.38854C10.3314 4.38817 10.3744 4.37913 10.4145 4.36194C10.4545 4.34475 10.4907 4.31976 10.521 4.28843C10.8698 3.94055 11.3423 3.74519 11.8349 3.74519C12.3275 3.74519 12.8 3.94055 13.1488 4.28843Z"
+                                  fill="#FB5B45"
+                                />
+                                <path
+                                  d="M6.12786 12.653C5.03039 12.653 3.95755 12.9784 3.04502 13.5881C2.13249 14.1978 1.42125 15.0644 1.00124 16.0783C0.581226 17.0922 0.471313 18.2079 0.685396 19.2843C0.89948 20.3607 1.42794 21.3494 2.20396 22.1255C2.97998 22.9015 3.96869 23.43 5.04507 23.6441C6.12146 23.8583 7.23716 23.7484 8.25109 23.3284C9.26503 22.9084 10.1317 22.1972 10.7414 21.2847C11.3511 20.3722 11.6765 19.2994 11.6765 18.2019C11.6749 16.7308 11.0898 15.3204 10.0496 14.2801C9.00936 13.2399 7.59898 12.6547 6.12786 12.653ZM6.12786 23.0945C5.16018 23.0946 4.21423 22.8076 3.40961 22.2701C2.605 21.7325 1.97787 20.9684 1.60753 20.0744C1.23719 19.1803 1.14027 18.1966 1.32904 17.2475C1.5178 16.2984 1.98376 15.4266 2.668 14.7424C3.35224 14.0581 4.22402 13.5921 5.1731 13.4033C6.12218 13.2145 7.10594 13.3114 7.99996 13.6817C8.89398 14.052 9.65811 14.6791 10.1957 15.4837C10.7333 16.2883 11.0203 17.2342 11.0203 18.2019C11.0186 19.4989 10.5026 20.7423 9.58544 21.6593C8.66828 22.5764 7.42485 23.0922 6.12786 23.0938V23.0945Z"
+                                  fill="#FB5B45"
+                                />
+                                <path
+                                  d="M8.1107 15.755L3.68101 20.1847C3.65054 20.2152 3.62637 20.2513 3.60988 20.2912C3.59339 20.331 3.5849 20.3736 3.5849 20.4167C3.5849 20.4598 3.59339 20.5025 3.60988 20.5423C3.62637 20.5821 3.65054 20.6183 3.68101 20.6488C3.71148 20.6792 3.74766 20.7034 3.78747 20.7199C3.82728 20.7364 3.86995 20.7449 3.91304 20.7449C3.95613 20.7449 3.9988 20.7364 4.03862 20.7199C4.07843 20.7034 4.1146 20.6792 4.14507 20.6488L8.57476 16.2191C8.6363 16.1575 8.67087 16.0741 8.67087 15.987C8.67087 15.9 8.6363 15.8165 8.57476 15.755C8.51322 15.6935 8.42976 15.6589 8.34273 15.6589C8.2557 15.6589 8.17224 15.6935 8.1107 15.755Z"
+                                  fill="#FB5B45"
+                                />
+                              </g>
+                              <defs>
+                                <clipPath id="clip0">
+                                  <rect
+                                    width="24"
+                                    height="24"
+                                    fill="white"
+                                    transform="translate(0.578979 0.3125)"
+                                  />
+                                </clipPath>
+                              </defs>
+                            </svg>
+                          </div>
+                          <div class="">
+                            <div class="title_1">Nifecard (30mg)</div>
+                            <div class="title_2">1 Pill (s); once Daily</div>
+                          </div>
+                        </div>
+                      </li>
+                      <li class="list-group-item">
+                        <div class="d-flex">
+                          <div class="mr-2">
+                            <svg
+                              width="25"
+                              height="25"
+                              viewBox="0 0 25 25"
+                              fill="none"
+                              xmlns="http://www.w3.org/2000/svg"
+                            >
+                              <g clip-path="url(#clip0)">
+                                <path
+                                  d="M23.1941 10.2749L15.1781 2.25913C14.2903 1.37761 13.0893 0.883886 11.8383 0.886101C10.5872 0.888315 9.38795 1.38629 8.50329 2.27095C7.61864 3.1556 7.12066 4.35482 7.11845 5.60591C7.11623 6.857 7.60996 8.05798 8.49148 8.94576L16.5075 16.9625C17.3942 17.8492 18.5968 18.3473 19.8508 18.3473C21.1048 18.3473 22.3075 17.8491 23.1942 16.9624C24.0809 16.0757 24.579 14.873 24.579 13.619C24.579 12.365 24.0808 11.1624 23.1941 10.2757V10.2749ZM7.76285 5.60245C7.76351 4.79737 8.00272 4.01055 8.45026 3.34133C8.89781 2.67212 9.53362 2.15053 10.2774 1.84243C11.0212 1.53433 11.8396 1.45355 12.6293 1.61027C13.4189 1.767 14.1445 2.15421 14.7142 2.72301L18.4903 6.49907L12.7314 12.2578L8.95554 8.4817C8.57621 8.10449 8.27548 7.65578 8.07075 7.16155C7.86603 6.66733 7.76137 6.1374 7.76285 5.60245ZM22.7302 16.4977C21.9658 17.26 20.9304 17.6881 19.8509 17.6881C18.7714 17.6881 17.7359 17.26 16.9715 16.4977L13.1955 12.7218L18.954 6.96295L22.7302 10.7388C23.4926 11.5032 23.9207 12.5387 23.9207 13.6183C23.9207 14.6978 23.4926 15.7333 22.7302 16.4977Z"
+                                  fill="#FB5B45"
+                                />
+                                <path
+                                  d="M13.1488 4.28843C13.1793 4.3189 13.2155 4.34307 13.2553 4.35956C13.2951 4.37605 13.3378 4.38454 13.3809 4.38454C13.424 4.38454 13.4666 4.37605 13.5064 4.35956C13.5462 4.34307 13.5824 4.3189 13.6129 4.28843C13.6434 4.25796 13.6675 4.22178 13.684 4.18197C13.7005 4.14216 13.709 4.09949 13.709 4.0564C13.709 4.01331 13.7005 3.97064 13.684 3.93082C13.6675 3.89101 13.6434 3.85484 13.6129 3.82437C13.141 3.35355 12.5016 3.08914 11.835 3.08914C11.1684 3.08914 10.529 3.35355 10.0571 3.82437C10.0258 3.85463 10.0008 3.89083 9.98357 3.93086C9.96637 3.97089 9.95731 4.01394 9.95692 4.05751C9.95653 4.10108 9.96483 4.14429 9.98132 4.18462C9.99781 4.22495 10.0222 4.26159 10.053 4.2924C10.0838 4.32322 10.1204 4.34759 10.1607 4.36409C10.201 4.3806 10.2443 4.38891 10.2878 4.38854C10.3314 4.38817 10.3744 4.37913 10.4145 4.36194C10.4545 4.34475 10.4907 4.31976 10.521 4.28843C10.8698 3.94055 11.3423 3.74519 11.8349 3.74519C12.3275 3.74519 12.8 3.94055 13.1488 4.28843Z"
+                                  fill="#FB5B45"
+                                />
+                                <path
+                                  d="M6.12786 12.653C5.03039 12.653 3.95755 12.9784 3.04502 13.5881C2.13249 14.1978 1.42125 15.0644 1.00124 16.0783C0.581226 17.0922 0.471313 18.2079 0.685396 19.2843C0.89948 20.3607 1.42794 21.3494 2.20396 22.1255C2.97998 22.9015 3.96869 23.43 5.04507 23.6441C6.12146 23.8583 7.23716 23.7484 8.25109 23.3284C9.26503 22.9084 10.1317 22.1972 10.7414 21.2847C11.3511 20.3722 11.6765 19.2994 11.6765 18.2019C11.6749 16.7308 11.0898 15.3204 10.0496 14.2801C9.00936 13.2399 7.59898 12.6547 6.12786 12.653ZM6.12786 23.0945C5.16018 23.0946 4.21423 22.8076 3.40961 22.2701C2.605 21.7325 1.97787 20.9684 1.60753 20.0744C1.23719 19.1803 1.14027 18.1966 1.32904 17.2475C1.5178 16.2984 1.98376 15.4266 2.668 14.7424C3.35224 14.0581 4.22402 13.5921 5.1731 13.4033C6.12218 13.2145 7.10594 13.3114 7.99996 13.6817C8.89398 14.052 9.65811 14.6791 10.1957 15.4837C10.7333 16.2883 11.0203 17.2342 11.0203 18.2019C11.0186 19.4989 10.5026 20.7423 9.58544 21.6593C8.66828 22.5764 7.42485 23.0922 6.12786 23.0938V23.0945Z"
+                                  fill="#FB5B45"
+                                />
+                                <path
+                                  d="M8.1107 15.755L3.68101 20.1847C3.65054 20.2152 3.62637 20.2513 3.60988 20.2912C3.59339 20.331 3.5849 20.3736 3.5849 20.4167C3.5849 20.4598 3.59339 20.5025 3.60988 20.5423C3.62637 20.5821 3.65054 20.6183 3.68101 20.6488C3.71148 20.6792 3.74766 20.7034 3.78747 20.7199C3.82728 20.7364 3.86995 20.7449 3.91304 20.7449C3.95613 20.7449 3.9988 20.7364 4.03862 20.7199C4.07843 20.7034 4.1146 20.6792 4.14507 20.6488L8.57476 16.2191C8.6363 16.1575 8.67087 16.0741 8.67087 15.987C8.67087 15.9 8.6363 15.8165 8.57476 15.755C8.51322 15.6935 8.42976 15.6589 8.34273 15.6589C8.2557 15.6589 8.17224 15.6935 8.1107 15.755Z"
+                                  fill="#FB5B45"
+                                />
+                              </g>
+                              <defs>
+                                <clipPath id="clip0">
+                                  <rect
+                                    width="24"
+                                    height="24"
+                                    fill="white"
+                                    transform="translate(0.578979 0.3125)"
+                                  />
+                                </clipPath>
+                              </defs>
+                            </svg>
+                          </div>
+                          <div class="">
+                            <div class="title_1">
+                              Vitamin C High Strength (1000mg)
+                            </div>
+                            <div class="title_2">
+                              Dissolve 1 Pill (s) in water; once daily; 5 days
+                            </div>
+                          </div>
+                        </div>
+                      </li>
+                    </ul>
+                    <div class="card-footer">
+                      <div class="title text-right">Manage Prescriptions</div>
+                    </div>
+                  </div>
+                </div>
+              </div>
+              <div class="col-md-4">
+                <div class="heading-title my-4 pl-4">Last Seen Doctors</div>
+                <div>
+                  <div class="card">
+                    <div class="">
+                      <div class="mx-4 py-3 b_b">
+                        <div class="d-flex mb-1">
+                          <div>
+                            <img
+                              class="rounded-img-sm mr-2"
+                              src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                            />
+                          </div>
+                          <div class="d-block">
+                            <p class="m-0 p-0 title_1">Aboki Joke</p>
+                            <p class="m-0 p-0 title_2">
+                              Obstetrics & Gynaecology
+                            </p>
+                          </div>
+                        </div>
+                        <div class="title_2">03 Aug; 3:10 PM</div>
+                        <div class="title_3 text-center mt-3 m-1">
+                          Schedule Appointment
+                        </div>
+                      </div>
+                      <div class="mx-4 py-3 b_b">
+                        <div class="d-flex mb-1">
+                          <div>
+                            <img
+                              class="rounded-img-sm mr-2"
+                              src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                            />
+                          </div>
+                          <div class="d-block">
+                            <p class="m-0 p-0 title_1">Dr Emeka Emeka Ike</p>
+                            <p class="m-0 p-0 title_2">General Surgeon</p>
+                          </div>
+                        </div>
+                        <div class="title_2">03 Aug; 3:10 PM</div>
+                        <div class="title_3 text-center mt-3 m-1">
+                          Schedule Appointment
+                        </div>
+                      </div>
+                    </div>
+                    <div class="card-footer">
+                      <div class="title text-right">See All</div>
+                    </div>
+                  </div>
+                </div>
+              </div>
+              <div class="col-md-4">
+                <div class="heading-title my-4 pl-4">Last Payment</div>
+                <div>
+                  <div class="card">
+                    <div class="card-body">
+                      <div class="row d-flex">
+                        <div class="col-md-6 col-sm-6">
+                          <p class="title_4 mb-1">Total Amount</p>
+                          <p class="title_5">&#8358; 23, 000</p>
+                        </div>
+                        <div class="col-md-6 col-sm-6">
+                          <p class="title_4 mb-1">Date:</p>
+                          <p class="title_1">03 Aug; 3:10 PM</p>
+                        </div>
+                      </div>
+
+                      <div class="d-flex mb-3">
+                        <div>
+                          <img
+                            class="rounded-img-sm mr-2"
+                            src="https://i.pinimg.com/564x/29/7b/11/297b11432a31c1e0c439bd97e86968b9.jpg"
+                          />
+                        </div>
+                        <div class="">
+                          <p class="m-0 p-0 title_1">Dr Shola Jimoh</p>
+                          <p class="m-0 p-0 title_2">
+                            Obstetrics & Gynaecology
+                          </p>
+                        </div>
+                      </div>
+
+                      <div class="">
+                        <div class="row">
+                          <div class="col-md-6">
+                            <span class="title_4"> Consultation </span>
+                          </div>
+                          <div class="col-md-6 text-right">
+                            <span class="title_2">3, 000: 00 </span>
+                          </div>
+                        </div>
+                      </div>
+
+                      <div class="row mt-2">
+                        <div class="col title_4 my-2">Medication</div>
+                      </div>
+
+                      <div class="d-flex justify-content-between align-items-center">
+                        <div>
+                          <p class="m-0 title_6">Paracetamol </p>
+                          <p class="m-0 title_7">
+                            1 Pill (s); thrice Daily; 3 days
+                          </p>
+                        </div>
+                        <div class="text-right">
+                          <p class="m-0 title_2">500: 00</p>
+                        </div>
+                      </div>
+
+                      <div class="d-flex justify-content-between align-items-center mt-3">
+                        <div>
+                          <p class="m-0 title_4">Tax</p>
+                        </div>
+                        <div class="text-right">
+                          <p class="m-0 title_2">500: 00</p>
+                        </div>
+                      </div>
+                    </div>
+                    <div class="card-footer">
+                      <div class="title text-right">View Payment History</div>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>
+      </div>
+    );
+  }
+}
