+++ src/Components/dashboard/SubHeader.js
@@ -1,0 +1,47 @@
+import React, { Component } from "react";
+
+export default class SubHeader extends Component {
+  render() {
+    return (
+      <div class="row main-top mb-4">
+        <div class="col-md-5 col-sm-12">
+          <h4>Welcome Joke,</h4>
+          <p class="title">
+            You have 3 notifications and an appointment scheduled for nextweek
+            with Dr. Shola Jimoh
+            <span>
+              <svg
+                width="25"
+                height="25"
+                viewBox="0 0 25 25"
+                fill="none"
+                xmlns="http://www.w3.org/2000/svg"
+              >
+                <g clip-path="url(#clip0)">
+                  <path
+                    d="M24.0647 14.0659C23.8873 12.8372 22.9221 11.8719 21.6934 11.6946V10.9674C21.6934 8.96103 20.1682 7.32746 18.2935 7.32746C16.4727 7.32746 14.9483 8.91143 14.8935 10.8586H14.9167L14.8935 10.8626V18.7453C14.8935 21.2653 13.0095 23.3116 10.6936 23.3116C8.37762 23.3116 6.49365 21.2625 6.49365 18.7453V18.0625C7.07364 18.0625 7.69363 17.6421 7.69363 16.4625V15.4565C10.2608 14.8001 12.0936 12.5958 12.0936 10.0966V9.26263H11.2936L11.2768 4.0567C11.271 3.09307 10.5795 2.27028 9.6312 2.09873C9.42665 1.47209 8.75281 1.1299 8.12612 1.33449C7.49948 1.53904 7.15729 2.21288 7.36189 2.83957C7.56643 3.46621 8.24027 3.80841 8.86696 3.60381C9.20251 3.49426 9.47235 3.24182 9.604 2.91432C10.1164 3.06012 10.4714 3.52631 10.4756 4.0591L10.4936 9.26263H9.6936V10.1082C9.6617 11.4995 8.72891 12.7086 7.39123 13.0926C6.6755 13.3197 5.90701 13.3197 5.19127 13.0926C3.85514 12.7073 2.92446 11.4984 2.89371 10.1082V9.26263H1.69372V4.0627C1.69602 3.52206 2.05962 3.04967 2.58171 2.90912C2.82516 3.52886 3.5249 3.83386 4.14464 3.59041C4.76438 3.34697 5.06937 2.64718 4.82593 2.02748C4.58248 1.40779 3.88269 1.10275 3.263 1.34619C2.92721 1.47814 2.66811 1.75349 2.55691 2.09673C1.59828 2.26053 0.896387 3.09017 0.893737 4.0627V9.26263H0.09375V10.0946C0.09375 12.5966 1.92772 14.8017 4.49368 15.4565V16.4625C4.49368 17.6421 5.11367 18.0625 5.69366 18.0625V18.7453C5.69366 21.7052 7.93643 24.1116 10.6936 24.1116C13.4507 24.1116 15.6935 21.7024 15.6935 18.7453V10.8818C15.7335 9.36383 16.9007 8.12865 18.2935 8.12865C19.727 8.12865 20.8934 9.40223 20.8934 10.9686V11.6946C19.3629 11.9155 18.3013 13.3353 18.5222 14.8658C18.7431 16.3964 20.1629 17.458 21.6934 17.2371C23.2239 17.0162 24.2856 15.5963 24.0647 14.0659ZM8.49362 2.86272C8.27272 2.86272 8.09362 2.68362 8.09362 2.46273C8.09362 2.24183 8.27272 2.06273 8.49362 2.06273C8.71451 2.06273 8.89361 2.24183 8.89361 2.46273C8.89361 2.68362 8.71451 2.86272 8.49362 2.86272ZM3.69369 2.06273C3.91459 2.06273 4.09369 2.24183 4.09369 2.46273C4.09369 2.68362 3.91459 2.86272 3.69369 2.86272C3.4728 2.86272 3.2937 2.68362 3.2937 2.46273C3.2937 2.24183 3.4728 2.06273 3.69369 2.06273ZM0.893737 10.0946V10.0626H2.09372V10.1082C2.12372 11.8467 3.28035 13.364 4.94887 13.8534C5.82321 14.1302 6.7617 14.1302 7.63603 13.8534C9.30536 13.3646 10.463 11.8473 10.4936 10.1082V10.0626H11.2936V10.0946C11.2936 12.3198 9.5784 14.2778 7.21364 14.7513C6.65835 14.8625 6.08956 14.8904 5.52606 14.8341C5.34057 14.8161 5.15617 14.7882 4.97367 14.7505C2.60891 14.2778 0.893737 12.3198 0.893737 10.0946ZM5.69366 17.2625C5.33367 17.2625 5.29367 16.7025 5.29367 16.4625V15.6137C5.32567 15.6177 5.35847 15.6193 5.39087 15.6225C5.43087 15.6269 5.47087 15.6301 5.51086 15.6337C5.70006 15.6505 5.89086 15.6617 6.08326 15.6621H6.10326C6.29525 15.6621 6.48565 15.6505 6.67365 15.6337C6.71365 15.6301 6.75605 15.6269 6.79724 15.6225C6.82884 15.6189 6.86164 15.6177 6.89324 15.6137V16.4625C6.89324 16.7025 6.85324 17.2625 6.49325 17.2625H5.69366ZM21.2934 16.4625C20.1889 16.4625 19.2934 15.5671 19.2934 14.4626C19.2934 13.358 20.1889 12.4626 21.2934 12.4626C22.3979 12.4626 23.2934 13.358 23.2934 14.4626C23.2921 15.5666 22.3974 16.4612 21.2934 16.4625Z"
+                    fill="#FB5B45"
+                  />
+                  <path
+                    d="M21.2937 13.2626C20.631 13.2626 20.0938 13.7998 20.0938 14.4626C20.0938 15.1253 20.631 15.6625 21.2937 15.6625C21.9565 15.6625 22.4937 15.1253 22.4937 14.4626C22.4937 13.7998 21.9565 13.2626 21.2937 13.2626ZM21.2937 14.8626C21.0728 14.8626 20.8937 14.6835 20.8937 14.4626C20.8937 14.2417 21.0728 14.0626 21.2937 14.0626C21.5146 14.0626 21.6937 14.2417 21.6937 14.4626C21.6937 14.6835 21.5146 14.8626 21.2937 14.8626Z"
+                    fill="#FB5B45"
+                  />
+                </g>
+                <defs>
+                  <clipPath id="clip0">
+                    <rect
+                      width="24"
+                      height="24"
+                      fill="white"
+                      transform="translate(0.09375 0.687012)"
+                    />
+                  </clipPath>
+                </defs>
+              </svg>
+            </span>
+          </p>
+        </div>
+      </div>
+    );
+  }
+}
