+++ src/Components/dashboard/index.css
@@ -9,7 +9,7 @@ body {
   margin-top: 24px !important;
 }
 
-.heading {
+#patient-dashboard .heading-title {
   font-family: "Roboto", sans-serif;
   font-style: normal;
   font-weight: 600;
@@ -33,7 +33,7 @@ body {
 
   color: #141515;
 }
-.card .title_1 {
+.title_1 {
   font-family: "Roboto", sans-serif;
   font-style: normal;
   font-weight: 500;
@@ -78,17 +78,13 @@ body {
   color: #706868;
 }
 
-.card .title_5 {
+.title_5 {
   font-family: Roboto;
   font-style: normal;
   font-weight: 600;
   font-size: 18px;
   line-height: 28px;
-  display: flex;
-  align-items: center;
-  text-align: center;
   letter-spacing: 0.02em;
-
   color: #141515;
 }
 
@@ -116,6 +112,14 @@ body {
   letter-spacing: 0.02em;
 
   color: #706868;
+}
+
+.card .title_8 {
+  font-family: Roboto;
+  font-size: 16px;
+  font-weight: normal;
+  color: #9e9e9e;
+  letter-spacing: 0.02em;
 }
 
 .card-header,
@@ -219,6 +223,7 @@ body {
   height: 60px !important;
   position: fixed;
   width: calc(100vw - 240px);
+  box-shadow: 0 4px 4px -4px rgba(0, 0, 0, 0.12);
 }
 
 .header-pd-lg {
@@ -229,9 +234,11 @@ body {
   height: 60px !important;
   position: fixed;
   width: 100vw;
+  box-shadow: 0 4px 4px -4px rgba(0, 0, 0, 0.12);
 }
 
-.header-pd.l-260, .header-pd-lg.l-260 {
+.header-pd.l-260,
+.header-pd-lg.l-260 {
   left: 240px;
   right: 0 !important;
 }
@@ -369,7 +376,7 @@ body {
 }
 
 .main-top .title {
-  font-family: Roboto;
+  font-family: "Roboto", sans-serif;
   font-style: normal;
   font-weight: normal;
   font-size: 12px;
@@ -379,4 +386,143 @@ body {
   letter-spacing: 0.02em;
   color: #706868;
   margin: 0;
+}
+
+/* profile */
+
+.profile-img-rounded {
+  border-radius: 50%;
+  width: 170px;
+  height: 170px;
+}
+
+/* profile */
+
+/* appointments */
+
+.btn-appointment {
+  font-family: "Roboto", sans-serif;
+  background: #1f8c93;
+  border-radius: 3px;
+  font-family: Roboto;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 14px;
+  line-height: 19px;
+  letter-spacing: 0.02em;
+
+  color: #ffffff;
+}
+
+
+.btn-save {
+  font-family: "Roboto", sans-serif;
+  background: #1f8c93;
+  border-radius: 3px;
+  font-family: Roboto;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 14px;
+  line-height: 19px;
+  letter-spacing: 0.02em;
+  padding-left: 26px;
+  padding-right: 26px;
+  color: #ffffff;
+  margin-bottom: 30px !important;
+}
+
+.btn-appointment_1 {
+  font-family: "Roboto", sans-serif;
+  background: #ffffff;
+  border-radius: 3px;
+  font-family: Roboto;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 14px;
+  line-height: 19px;
+  letter-spacing: 0.02em;
+
+  color: #141515;
+}
+
+.title_md {
+  font-family: "Roboto", sans-serif;
+  font-style: bold;
+  font-weight: 600;
+  font-size: 18px;
+  line-height: 28px;
+  letter-spacing: 0.02em;
+  color: #141515;
+}
+
+/* end of appointments */
+
+.vl {
+  background: #fff;
+  padding: 16px !important;
+  position: relative;
+}
+
+#appointment thead {
+  background: #fff;
+}
+
+#appointment form{
+  position: relative;
+}
+
+.search-container .input {
+  background: #ffffff;
+  border: 1px solid rgba(0, 0, 0, 0.14);
+  margin: 0;
+  color: rgba(0, 0, 0, 0.87);
+  word-wrap: break-word;
+  outline: none;
+  display: flex;
+  flex: 100%;
+  -webkit-tap-highlight-color: transparent;
+  /*  margin-top: -37px; */
+  height: calc(1.5em + 0.75rem + 2px);
+  font-size: 16px;
+  padding: 0.375rem 1rem 0.375rem 0.75rem;
+  border-radius: 10px !important;
+}
+
+.search-icon {
+  position: absolute;
+  right: 5px;
+  height: calc(1.5em + 0.75rem + 2px);
+  border-radius: 30px !important;
+}
+
+#appointment search {
+  display: flex;
+  flex: 1;
+  flex-wrap: wrap;
+  height: calc(1.5em + 0.75rem + 2px);
+}
+
+ul {
+  text-decoration: none !important;
+  list-style: none !important;
+  padding-left: 0 !important;
+}
+
+#appointment .d3ext {
+  font-family: "Roboto", sans-serif;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 16px;
+  line-height: 19px;
+  letter-spacing: 0.02em;
+
+  color: #1f8c93;
+}
+
+@media (max-width: 768px) {
+  .main {
+    padding-left: 16px !important;
+    padding-right: 16px !important;
+    padding-top: 25px !important;
+  }
 }
